### 2022-09-24 #4791 @editorialbot
[PRE REVIEW]: localcovid19now: processing and mapping COVID-19 case data at subnational scales
**Submitting author:** @sjbeckett (<a href="http://orcid.org/0000-0002-4410-2960">Stephen Beckett</a>)
**Repository:** https://github.com/sjbeckett/localcovid19now
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.0.9900
