+++ paper/refs.bib
@@ -51,6 +51,7 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   number={5},
   pages={460--471},
   year={2020},
+  doi={10.1038/s41562-020-0884-z},
   publisher={Nature Publishing Group}
 }
 
@@ -62,6 +63,7 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   number={50},
   pages={1247--1256},
   year={2010},
+  doi={10.1098/rsif.2010.0142},
   publisher={The Royal Society}
 }
 
@@ -71,6 +73,7 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   author={Barber, Ryan M and Sorensen, Reed JD and Pigott, David M and Bisignano, Catherine and Carter, Austin and Amlag, Joanne O and Collins, James K and Abbafati, Cristiana and Adolph, Christopher and Allorant, Adrien and others},
   journal={The Lancet},
   year={2022},
+  doi={10.1016/S0140-6736(22)00484-6},
   publisher={Elsevier}
 }
 
@@ -82,6 +85,7 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   number={1},
   pages={1--10},
   year={2021},
+  doi={10.1038/s41597-021-00955-2},
   publisher={Nature Publishing Group}
 }
 
@@ -94,6 +98,7 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   number={51},
   pages={32764--32771},
   year={2020},
+  doi={10.1073/pnas.2009911117},
   publisher={National Acad Sciences}
 }
 
@@ -108,6 +113,7 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   number={12},
   pages={1313--1319},
   year={2020},
+  doi={10.1038/s41562-020-01000-9},
   publisher={Nature Publishing Group}
 }
 
@@ -118,7 +124,8 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   volume={6},
   number={63},
   pages={3290},
-  year={2021}
+  year={2021},
+  doi={10.21105/joss.03290}
 }
 
 
@@ -143,6 +150,7 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   number={8},
   pages={677--683},
   year={2021},
+  doi={10.1038/s43587-021-00095-7},
   publisher={Nature Publishing Group}
 }
 
@@ -152,7 +160,8 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   year={2022},
   publisher={PsyArXiv},
   journal={PsyArXiv},
-  URL ={https://doi.org/10.31234/osf.io/v8tdf}
+  URL ={https://doi.org/10.31234/osf.io/v8tdf},
+  doi={10.31234/osf.io/v8tdf}
 }
 
 
@@ -178,6 +187,7 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   number={5},
   pages={533--534},
   year={2020},
+  doi={10.1016/S1473-3099(20)30120-1},
   publisher={Elsevier}
 }
 
@@ -187,7 +197,8 @@ howpublished = "(accessed September 8, 2022) \url{https://www.bbc.com/news/world
   author={Cota, Wesley},
   year={2020},
   journal={SciELO Preprints},
-  URL = {https://doi.org/10.1590/SciELOPreprints.362}
[…]
