### 2022-11-07 #4898 @galessiorob
@arfon ran into the same issue with editorialbot, not sure what's causing the glitch.
