### 2022-11-15 sjbeckett/localcovid19now#58 @welch16
Having issues installing with devtools::install_github
Hello,

Great work with the package. I am reviewing the package for JOSS, and when I am trying to install the package using the `devtools::install_github` instruction, I got the following error:

```
Downloading GitHub repo sjbeckett/localcovid19now@HEAD
Error in utils::download.file(url, path, method = method, quiet = quiet,  :
  download from 'https://api.github.com/repos/sjbeckett/localcovid19now/tarball/HEAD' failed
```

However, if download/clone the repository and use `devtools::install_local` then it is possible to install the package. I suspect that this issue may happen to other users, in which case may be useful to have alternative instructions to install the package
