### 2022-11-15 #4898 @welch16
I experienced a minor issue installing the package using `devtools::install_github`. I added the following issue:

https://github.com/sjbeckett/localcovid19now/issues/58
