### 2022-11-23 sjbeckett/localcovid19now#58 @sjbeckett
Many thanks for reporting this @welch16. I'm glad the `devtools::install_local` method allowed you to install the package.
I'm unable to reproduce this error, and not exactly sure what the underlying cause is. I'm wondering if you could try again and see if you get the same error? and whether you have seen this with other packages that use the `devtools::install_github` approach?

If not, then it is possible that changing timeout options could help: https://github.com/dmcable/spacexr/issues/62
or changing download options could help: https://github.com/facebookincubator/GeoLift/issues/53
