### 2022-11-23 58c097a
This Commit 1) updated @description in mapping files 2) updated references to the manuscript in refs.bib
update note on google credentials

Updated Italy to pull dates and files from latest file, rather than manually asserting latest date.

Update New Zealand, using github api to find list of files in the NZ github repo to get the latest data.

update loading options for vroom calls to show_col_types = FALSE

fixed warning messages/ and removed other print outs to console.

silenced googledrive console output within LoadPhilippines

updated @ description in MakeTable.R

Update documentation an
[…]
