+++ R/LoadZimbabwe.R
@@ -8,15 +8,14 @@
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
-#' \dontrun{
 #' Zimbabwe <- LoadZimbabwe()
-#' }
 #' @seealso [LoadCountries()]
 #' @export
 LoadZimbabwe <- function() {
   # COVID-19 Data Repository by African Surveyors Connect https://github.com/African-Surveyors-Connect/Zimbabwe-COVID-19-Data/
   # Dashboard: https://surveyor-jr.maps.arcgis.com/apps/dashboards/8ef907d2658c44c6a143819aa7979b20
 
+  geomZimbabwe <- pop_zimbabwe <- NULL
   # note could be improved by calling data from ESRI, provincial timeseries feature Id: 20703dd3a24f45f08ea37034285d3492
   utils::data("geomZimbabwe", envir = environment())
   utils::data("pop_zimbabwe", envir = environment())
