+++ R/LoadPhilippines.R
@@ -14,13 +14,16 @@
 #' @seealso [LoadCountries()]
 #' @export
 LoadPhilippines <- function() {
-  utils::data("geomPhilippines", envir = environment())
+  . <- geomPeru <- pop_philippines <- name <- id <- ProvRes <- DateRepConf <- RegionRes <- TotalReported <- Province <- pop_philippines <- NULL
+  utils::data("geomPeru", envir = environment())
   utils::data("pop_philippines", envir = environment())
 
   # Republic of Philippines Department of Health: https://doh.gov.ph/covid19tracker
-  
+
   rlang::check_installed(c("pdftools", "googledrive"), reason = "to use `LoadPhilippines()`")
 
+  # silence googledrive in this specific scope
+  googledrive::local_drive_quiet()
 
   url1 <- "bit.ly/DataDropPH"
   req1 <- httr::GET(url1)
@@ -62,7 +65,7 @@ LoadPhilippines <- function() {
           DateDied = vroom::col_date(format = "%Y-%m-%d"),
           DateRecover = vroom::col_date(format = "%Y-%m-%d"),
           DateOnset = vroom::col_date(format = "%Y-%m-%d")
-        ), 
+        ),
         show_col_types = FALSE
       )
       unlink(temp)
