+++ R/LoadItaly.R
@@ -1,16 +1,3 @@
-#' dataQueryItaly
-#'
-#' @param date search for Italian COVID-19 data on this date
-#'
-#' @return returns data for Italy on specified date. Called by LoadItaly().
-#' @keywords internal
-dataQueryItaly <- function(date) {
-  denominazione_regione <- denominazione_provincia <- codice_provincia <- totale_casi <- NULL
-  data <- vroom::vroom(paste0("https://raw.githubusercontent.com/pcm-dpc/COVID-19/master/dati-province/dpc-covid19-ita-province-", stringr::str_replace_all(as.character(date), "-", ""), ".csv"), col_types = vroom::cols(note = vroom::col_character()), show_col_types = FALSE, progress = FALSE) %>%
-    dplyr::select(date = data, region = denominazione_regione, province = denominazione_provincia, code = codice_provincia, cases = totale_casi)
-  return(data)
-}
-
 #' LoadItaly
 #'
 #' @description Reads in subnational data for Italy to calculate most recent estimate of per capita active COVID-19 cases.
@@ -21,13 +8,11 @@ dataQueryItaly <- function(date) {
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
-#' \dontrun{
 #' Italy <- LoadItaly()
-#' }
 #' @seealso [LoadCountries()]
 #' @export
 LoadItaly <- function() {
-  code <- cases <- region <- province <- pop_italy <- NULL
+  code <- cases <- region <- province <- pop_italy <- totale_casi <- codice_provincia <- denominazione_regione <- denominazione_provincia <- NULL
 
   # Load geometry and population data
   utils::data("geomItaly", envir = environment())
@@ -36,44 +21,22 @@ LoadItaly <- function() {
 
   # Italian Department of Civil Protection COVID-19 Data: https://github.com/pcm-dpc/COVID-19/
   # italy: need to download data_cur and data_past respectively
-  cur_date <- lubridate::ymd(gsub("-", "", Sys.Date())) - 1
-  # past_date <- ymd(cur_date) - 14
-  #
-  # data_past <- dataQueryItaly(past_date) %>%
-  #   select(date, code, cases) # date, abbreviation_canton_and_fl, ncumul_conf
-  data_cur <- dataQueryItaly(cur_date)
-  for (i in c(1:13)) {
-    data_cur <- data_cur %>% rbind(dataQueryItaly(cur_date - i))
-  }
-  data_cur <- data_cur %>%
-    dplyr::group_by(code) %>%
-    dplyr::summarise(date = dplyr::first(date), cases = dplyr::first(cases), region = dplyr::first(region), province = dplyr::first(province), n = dplyr::n())
 
-  # This will actually give results inclusive of missing data
-  past_date <- unique(lubridate::as_date(data_cur$date)) - 14 # get all dates 14 days before a most recent current date
+  latestdata <- vroom::vroom(paste0("https://raw.githubusercontent.com/pcm-dpc/COVID-19/master/dati-province/dpc-covid19-ita-province-latest.csv"), col_types = vroom::cols(note = vroom::col_character())) %>%
+    dplyr::select(latest_date = data, region = denominazione_regione, province = denominazione_provincia, code = codice_provincia, cases = totale_casi)
 
-  data_past <- purrr::map_df(past_date, ~ dataQueryItaly(.x)) %>%
-    dplyr::select(date, code, cases) %>%
[…]
