+++ R/LoadNigeria.R
@@ -9,17 +9,15 @@
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
-#' \dontrun{
 #' Nigeria <- LoadNigeria()
-#' }
 #' @seealso [LoadCountries()]
 #' @export
 LoadNigeria <- function() {
-
   # Humanitarian Emergency Response Africa  https://data.humdata.org/dataset/nigeria_covid19_subnational
   # sourced from data collected by Nigeria Centre for Disease Control and National Primary Health Care Development Agency.
   # updated ~ weekly.
 
+  geomNigeria <- NULL
   utils::data("geomNigeria", envir = environment())
 
   casedata <- vroom::vroom("https://data.humdata.org/dataset/f5c35452-d766-468a-a272-4bd82d0a3be0/resource/15b9978f-422b-4b3d-9513-359b520d8352/download/nga_subnational_covid19_hera.csv", delim = ";", show_col_types = FALSE, progress = FALSE)
