+++ R/LoadSaudiArabia.R
@@ -18,7 +18,11 @@ LoadSaudiArabia <- function() {
 
   utils::data(list = c("misc_saudiarabia", "pop_saudiarabia", "geomSaudiArabia"), envir = environment())
 
+<<<<<<< HEAD
   Casesdataset <- sf::st_read("https://services6.arcgis.com/bKYAIlQgwHslVRaK/arcgis/rest/services/VWPlacesCasesHostedView/FeatureServer/0/query?where=1%3D1&outFields=*&outSR=4326&f=json", quiet = TRUE)
+=======
+  Casesdataset <- sf::st_read("https://services6.arcgis.com/bKYAIlQgwHslVRaK/arcgis/rest/services/VWPlacesCasesHostedView/FeatureServer/0/query?where=1%3D1&outFields=*&outSR=4326&f=json", quiet=TRUE)
+>>>>>>> 9cf4ba0 (fixed warning messages/ and removed other print outs to console.)
   # note need to convert dates from UNIX timestamp
 
   # 13 regions, 146 governates.
