+++ R/MakeTable.R
@@ -1,5 +1,7 @@
 #' Calculate Risk
 #'
+#' @description Calculates the percentage probability that one or more persons in a group, of a particular size g, may be infectious given the underlying prevalence of disease.
+#'
 #' @param p_I Probability one individual in a population is infectious.
 #' @param g Event size.
 #'
@@ -11,12 +13,14 @@
 #' risk <- calcrisk(.001, 50)
 #' }
 #'
-calc_risk <- function(p_I, g) {
+calcRisk <- function(p_I, g) {
   r <- 1 - (1 - p_I)**g
   return(r * 100)
 }
 
 #' Create Table of Risk Estimates
+#'
+#' @description Creates a table showing the estimated risk that one or more people will be infectious for the given input locations, event sizes and ascertainment biases.
 #'
 #' @param df_in Input data.
 #' @param risk_output Name of output file.
@@ -75,7 +79,7 @@ create_c19r_data <- function(df_in,
 
       riskdt <- data_Nr %>%
         dplyr::mutate(
-          risk = round(calc_risk(
+          risk = round(calcRisk(
             Nr, size
           ), 0),
           risk = dplyr::case_when(
@@ -91,7 +95,7 @@ create_c19r_data <- function(df_in,
     }
   }
 
-  risk_data_df <- purrr::reduce(.x = append(list(df_in), risk_data), .f = dplyr::left_join) %>%
+  risk_data_df <- purrr::reduce(.x = append(list(df_in), risk_data), .f = dplyr::left_join, by="geoid") %>%
     dplyr::mutate(updated = lubridate::ymd(gsub("-", "", Sys.Date())))
 
   utils::write.csv(risk_data_df,
