+++ R/LoadItaly.R
@@ -1,16 +1,3 @@
-#' dataQueryItaly
-#'
-#' @param date search for Italian COVID-19 data on this date
-#'
-#' @return returns data for Italy on specified date. Called by LoadItaly().
-#' @keywords internal
-dataQueryItaly <- function(date) {
-  denominazione_regione <- denominazione_provincia <- codice_provincia <- totale_casi <- NULL
-  data <- vroom::vroom(paste0("https://raw.githubusercontent.com/pcm-dpc/COVID-19/master/dati-province/dpc-covid19-ita-province-", stringr::str_replace_all(as.character(date), "-", ""), ".csv"), col_types = vroom::cols(note = vroom::col_character())) %>%
-    dplyr::select(date = data, region = denominazione_regione, province = denominazione_provincia, code = codice_provincia, cases = totale_casi)
-  return(data)
-}
-
 #' LoadItaly
 #'
 #' @description Reads in subnational data for Italy to calculate most recent estimate of per capita active COVID-19 cases.
@@ -36,45 +23,23 @@ LoadItaly <- function() {
 
   # Italian Department of Civil Protection COVID-19 Data: https://github.com/pcm-dpc/COVID-19/
   # italy: need to download data_cur and data_past respectively
-  cur_date <- lubridate::ymd(gsub("-", "", Sys.Date())) - 1
-  # past_date <- ymd(cur_date) - 14
-  #
-  # data_past <- dataQueryItaly(past_date) %>%
-  #   select(date, code, cases) # date, abbreviation_canton_and_fl, ncumul_conf
-  data_cur <- dataQueryItaly(cur_date)
-  for (i in c(1:13)) {
-    data_cur <- data_cur %>% rbind(dataQueryItaly(cur_date - i))
-  }
-  data_cur <- data_cur %>%
-    dplyr::group_by(code) %>%
-    dplyr::summarise(date = dplyr::first(date), cases = dplyr::first(cases), region = dplyr::first(region), province = dplyr::first(province), n = dplyr::n())
-
-  # This will actually give results inclusive of missing data
-  past_date <- unique(lubridate::as_date(data_cur$date)) - 14 # get all dates 14 days before a most recent current date
-
-  data_past <- purrr::map_df(past_date, ~ dataQueryItaly(.x)) %>%
-    dplyr::select(date, code, cases) %>%
-    dplyr::inner_join( # only keep the past dates that are 14 days before their corresponding date
-      data_cur %>%
-        dplyr::select(code, date) %>%
-        dplyr::mutate(date = date - lubridate::period(14, "days"))
-    )
-
-
-  # population
-  # pop <- vroom("countries/data/italyPop.csv", col_types = cols(code = col_character())) %>%
-  #   mutate(
-  #     code = str_pad(code, width = 3, side = "left", pad = "0")
-  #   )
-  # data("pop_italy")
-
-  data_join <- data_cur %>%
-    dplyr::inner_join(data_past, by = "code", suffix = c("", "_past")) %>%
+    
+    latestdata <- vroom::vroom(paste0("https://raw.githubusercontent.com/pcm-dpc/COVID-19/master/dati-province/dpc-covid19-ita-province-latest.csv"), col_types = vroom::cols(note = vroom::col_character())) %>%
+    dplyr::select(latest_date = data, region = denominazione_regione, province = denominazione_provincia, code = codice_provincia, cases = totale_casi)
+  
+    pastdate = lubridate::date(latestdata$latest_date[1]) - lubridate::days(14)
+  
+    
[…]
