### 2022-11-23 de83867
Update New Zealand, using github api to find list of files in the NZ github repo to get the latest data.
