+++ R/LoadMexico.R
@@ -27,7 +27,7 @@ LoadMexico <- function() {
     DATE <- Sys.Date() - aa
     formDATE <- format(DATE, "%Y%m%d")
     STRING <- paste0("https://datos.covid-19.conacyt.mx/Downloads/Files/Casos_Diarios_Municipio_Confirmados_", formDATE)
-    MEX <- try(vroom::vroom(STRING)) # note older files are DELETED.
+    MEX <- try(vroom::vroom(STRING),show_col_types=FALSE) # note older files are DELETED.
     if (is.null(dim(MEX)) == FALSE) {
       flag <- 1
     } else {
