+++ R/LoadBelgium.R
@@ -41,7 +41,7 @@ LoadBelgium <- function() {
     STRING <- paste0("https://epistat.sciensano.be/Data/", getDate(aa, 0), "/COVID19BE_CASES_MUNI_CUM_", getDate(aa, 0), ".csv")
     tryCatch(
       {
-        latest_data <- vroom::vroom(STRING)
+        latest_data <- vroom::vroom(STRING, show_col_types=FALSE)
       },
       error = function(cond) {
         warning(paste0("No data for ", getDate(aa, 0)))
