+++ R/LoadNorway.R
@@ -16,7 +16,7 @@
 LoadNorway <- function() {
   utils::data("geomNorway", envir = environment())
 
-  data <- vroom::vroom("https://raw.githubusercontent.com/thohan88/covid19-nor-data/master/data/01_infected/msis/municipality.csv")
+  data <- vroom::vroom("https://raw.githubusercontent.com/thohan88/covid19-nor-data/master/data/01_infected/msis/municipality.csv", show_col_types=FALSE)
   data$date <- as.Date(data$date)
 
   ## THIS SCRIPT ASSUMES ALL COUNTIES UPDATE AT ONCE.
