+++ R/LoadVenezuela.R
@@ -18,7 +18,7 @@ LoadVenezuela <- function() {
 
   utils::data("geomVenezuela", envir = environment())
 
-  casedata <- vroom::vroom("https://docs.google.com/spreadsheets/d/e/2PACX-1vQI4s0no2TS1dYxbv82nhKD7iz8fbDGwdsOI4kzJ0cg3gjOR51KIw_rNOff97Xic_fRQD41xmsDGUfM/pub?gid=1029482781&single=true&output=csv")
+  casedata <- vroom::vroom("https://docs.google.com/spreadsheets/d/e/2PACX-1vQI4s0no2TS1dYxbv82nhKD7iz8fbDGwdsOI4kzJ0cg3gjOR51KIw_rNOff97Xic_fRQD41xmsDGUfM/pub?gid=1029482781&single=true&output=csv", show_col_types=FALSE)
 
   regions <- names(casedata)[1:25]
   DateReport <- rep(utils::tail(casedata$date, 1), length(regions))
