+++ R/LoadGhana.R
@@ -22,7 +22,7 @@ LoadGhana <- function() {
   geomGhana <- sf::st_as_sf(geomGhana)
 
   temp <- tempfile()
-  utils::download.file(url = "https://www.dropbox.com/s/2uxzix4upet0nlm/cases_ghana.csv?dl=1", destfile = temp)
+  utils::download.file(url = "https://www.dropbox.com/s/2uxzix4upet0nlm/cases_ghana.csv?dl=1", destfile = temp, quiet = TRUE)
   casesGhana <- vroom::vroom(temp, show_col_types=FALSE)
 
   ghanaCases <- casesGhana %>%
