+++ R/LoadIndonesia.R
@@ -15,7 +15,7 @@
 #' @seealso [LoadCountries()]
 #' @export
 LoadIndonesia <- function() {
-  cases <- vroom::vroom("https://raw.githubusercontent.com/erlange/INACOVID/master/data/csv/ext.prov.csv") # cases are KASUS. new cases per day.
+  cases <- vroom::vroom("https://raw.githubusercontent.com/erlange/INACOVID/master/data/csv/ext.prov.csv", show_col_types=FALSE) # cases are KASUS. new cases per day.
 
   utils::data("geomIndonesia", envir = environment())
 
