+++ R/MakeTable.R
@@ -1,5 +1,7 @@
 #' Calculate Risk
 #'
+#' @description Calculates the percentage probability that one or more persons in a group, of a particular size g, may be infectious given the underlying prevalence of disease.
+#'
 #' @param p_I Probability one individual in a population is infectious.
 #' @param g Event size.
 #'
@@ -17,6 +19,8 @@ calc_risk <- function(p_I, g) {
 }
 
 #' Create Table of Risk Estimates
+#'
+#' @description Creates a table showing the estimated risk that one or more people will be infectious for the given input locations, event sizes and ascertainment biases.
 #'
 #' @param df_in Input data.
 #' @param risk_output Name of output file.
