### 2022-11-23 cd89852
silenced joining message in create_c19r_data
