### 2022-11-25 sjbeckett/localcovid19now#59 @mbkumar
Running the example code leads to errors
I could not run this line. My machine has Centos-8 Stream.
```
> MAP = EventMap_tmap(US,100,US$AB,projection="ESPG:5070")
Error in CPL_transform(x, crs, aoi, pipeline, reverse, desired_accuracy,  :
  crs not found: is it missing?```
