### 2022-11-25 sjbeckett/localcovid19now#59 @mbkumar
I'd also like to bring the following warnings to your attention. I don't know if they are important or not.

```
> EventMap_tmap(US,100,US$AB)
Warning messages:
1: In CPL_transform(x, crs, aoi, pipeline, reverse, desired_accuracy,  :
  GDAL Error 1: PROJ: proj_as_wkt: Unsupported conversion method: Equal Earth
2: In CPL_transform(x, crs, aoi, pipeline, reverse, desired_accuracy,  :
  GDAL Error 1: PROJ: proj_as_wkt: Unsupported conversion method: Equal Earth
```
