### 2022-11-26 7db731c
Silence chatty functions
silenced joining message in create_c19r_data
