+++ R/LoadIreland.R
@@ -25,7 +25,7 @@ LoadIreland <- function() {
   # geom <<- st_read('https://raw.githubusercontent.com/appliedbinf/covid19-event-risk-planner/master/COVID19-Event-Risk-Planner/map_data/Ireland_Counties.geojson')
   # data("geomIreland")
   # Main COVID-19 hub page: https://covid-19.geohive.ie/datasets/d9be85b30d7748b5b7c09450b8aede63_0
-  data <- vroom::vroom("https://opendata.arcgis.com/datasets/d9be85b30d7748b5b7c09450b8aede63_0.csv") %>%
+  data <- vroom::vroom("https://opendata.arcgis.com/datasets/d9be85b30d7748b5b7c09450b8aede63_0.csv", show_col_types = FALSE, progress = FALSE) %>%
     dplyr::mutate(date = as.Date(TimeStamp)) %>%
     dplyr::select(CountyName, date, cases = ConfirmedCovidCases, pop = PopulationCensus16) %>%
     dplyr::arrange(dplyr::desc(date))
