+++ R/LoadMalaysia.R
@@ -19,7 +19,7 @@ LoadMalaysia <- function() {
   utils::data("geomMalaysia", envir = environment())
   utils::data("pop_malaysia", envir = environment())
 
-  casesbystate <- vroom::vroom("https://raw.githubusercontent.com/MoH-Malaysia/covid19-public/main/epidemic/cases_state.csv") # new cases by state by time
+  casesbystate <- vroom::vroom("https://raw.githubusercontent.com/MoH-Malaysia/covid19-public/main/epidemic/cases_state.csv", show_col_types = FALSE, progress = FALSE) # new cases by state by time
   casesbystate$date <- lubridate::as_date(casesbystate$date)
   states <- unique(casesbystate$state)
 

[…]
