+++ R/LoadGhana.R
@@ -22,8 +22,8 @@ LoadGhana <- function() {
   geomGhana <- sf::st_as_sf(geomGhana)
 
   temp <- tempfile()
-  utils::download.file(url = "https://www.dropbox.com/s/2uxzix4upet0nlm/cases_ghana.csv?dl=1", destfile = temp)
-  casesGhana <- vroom::vroom(temp)
+  utils::download.file(url = "https://www.dropbox.com/s/2uxzix4upet0nlm/cases_ghana.csv?dl=1", destfile = temp, quiet = TRUE)
+  casesGhana <- vroom::vroom(temp, show_col_types = FALSE, progress = FALSE)
 
   ghanaCases <- casesGhana %>%
     dplyr::group_by(name) %>%
