### 2022-11-26 54c0df8
RCMDCHECK and documentation updates
