+++ R/LoadCountries.R
@@ -8,9 +8,7 @@
 #' @export
 #'
 #' @examples
-#' \dontrun{
 #' AllData <- LoadCountries()
-#' }
 LoadCountries <- function(countries = NULL, filepath = NULL, interactiveMode = interactive()) {
 
   # COMBINE DATASETS INTO SINGLE OBJECT
