+++ R/LoadPakistan.R
@@ -11,14 +11,13 @@
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
-#' \dontrun{
 #' Pakistan <- LoadPakistan()
-#' }
 #' @seealso [LoadCountries()]
 #' @export
 LoadPakistan <- function() {
   # Data obtained from COVID-19 Data Hub https://covid19datahub.io
 
+  geomPakistan <- NULL
   utils::data("geomPakistan", envir = environment())
 
   # Guidotti et al., (2020). COVID-19 Data Hub. Journal of Open Source Software, 5(51), 2376, https://doi.org/10.21105/joss.02376
