### 2022-11-26 4bea343
RCMDCHECK and documentation updates
