+++ R/LoadItaly.R
@@ -21,13 +21,11 @@ dataQueryItaly <- function(date) {
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
-#' \dontrun{
 #' Italy <- LoadItaly()
-#' }
 #' @seealso [LoadCountries()]
 #' @export
 LoadItaly <- function() {
-  code <- cases <- region <- province <- pop_italy <- NULL
+  code <- cases <- region <- province <- pop_italy <- totale_casi <- codice_provincia <- denominazione_regione <- denominazione_provincia <- NULL
 
   # Load geometry and population data
   utils::data("geomItaly", envir = environment())
