+++ R/LoadIndonesia.R
@@ -9,14 +9,13 @@
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
-#' \dontrun{
 #' Indonesia <- LoadIndonesia()
-#' }
 #' @seealso [LoadCountries()]
 #' @export
 LoadIndonesia <- function() {
   cases <- vroom::vroom("https://raw.githubusercontent.com/erlange/INACOVID/master/data/csv/ext.prov.csv", show_col_types = FALSE, progress = FALSE) # cases are KASUS. new cases per day.
 
+  micro_name <- NULL
   utils::data("geomIndonesia", envir = environment())
 
   provinces <- unique(cases$Location)
