+++ R/LoadUS.R
@@ -9,9 +9,7 @@
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
-#' \dontrun{
 #' US <- LoadUS()
-#' }
 #' @seealso [LoadCountries()]
 #' @export
 LoadUS <- function() {
