+++ R/addNewGeom.R
@@ -23,10 +23,11 @@ moveFiles <- function(x) {
 #' @export
 #'
 #' @examples
-#' \dontrun{
 #' addNewGeoms()
-#' }
 addNewGeoms <- function() {
+  
+  iso3 <- m49code <- macro_code <- micro_code <- geoid <- country_name <- macro_name <- NULL
+  
   rlang::check_installed(c("here"), reason = "to use addNewGeoms()")
   # Fields will be as follows:
   ## geoid = ISO3m49_micro_macro
@@ -81,9 +82,7 @@ addNewGeoms <- function() {
 #' @export
 #'
 #' @examples
-#' \dontrun{
 #' resetNewGeoms()
-#' }
 resetNewGeoms <- function() {
   rlang::check_installed(c("here"), reason = "to use resetNewGeoms()")
   file_list <- list.files(here::here("tools", "processed"))
