### 2022-11-27 6a3b0f4
Remove unused dataQueryItaly function
Signed-off-by: Aroon Chande <github@atc.io>

remove deprecated function makeMaps.R

case correction in example for calcRisk

case correction for calcRisk

corrected pointing to missing data files
