+++ man/makeMaps.Rd
@@ -1,15 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/MakeMaps.R
-\name{makeMaps}
-\alias{makeMaps}
-\title{makeMaps}
-\usage{
-makeMaps()
-}
-\description{
-Creates a whole world risk map and saves to PNG
-}
-\examples{
-makeMaps()
-
-}
