+++ NAMESPACE
@@ -57,7 +57,6 @@ export(addNewGeoms)
 export(calcRisk)
 export(create_c19r_data)
 export(estRisk)
-export(makeMaps)
 export(moveFiles)
 export(remSurplus)
 export(resetNewGeoms)
