### 2022-11-27 cafa10e
Remove unused dataQueryItaly function
Signed-off-by: Aroon Chande <github@atc.io>
