+++ R/LoadVietnam.R
@@ -25,7 +25,7 @@ LoadVietnam <- function() {
   # ProvinceInfo=data.frame(ProvinceInfo)
   # write.csv(ProvinceInfo,"countries/data/VietnamProvinceInfo.csv",row.names = FALSE)
   # ProvinceInfo <- read.csv("countries/data/VietnamProvinceInfo.csv")
-  
+
   misc_vietnam <- geomVietnam <- NULL
   utils::data("misc_vietnam", envir = environment())
   utils::data("geomVietnam", envir = environment())

[…]
