+++ R/LoadCountries.R
@@ -10,7 +10,6 @@
 #' @examples
 #' AllData <- LoadCountries()
 LoadCountries <- function(countries = NULL, filepath = NULL, interactiveMode = interactive()) {
-
   # COMBINE DATASETS INTO SINGLE OBJECT
   NEWMAP <- c()
 
