+++ R/LoadPhilippines.R
@@ -12,17 +12,16 @@
 #' @seealso [LoadCountries()]
 #' @export
 LoadPhilippines <- function() {
-  
   . <- geomPeru <- pop_philippines <- name <- id <- ProvRes <- DateRepConf <- RegionRes <- TotalReported <- Province <- pop_philippines <- NULL
   utils::data("geomPeru", envir = environment())
   utils::data("pop_philippines", envir = environment())
 
   # Republic of Philippines Department of Health: https://doh.gov.ph/covid19tracker
-  
+
   rlang::check_installed(c("pdftools", "googledrive"), reason = "to use `LoadPhilippines()`")
-  
+
   # silence googledrive in this specific scope
-  googledrive::local_drive_quiet() 
+  googledrive::local_drive_quiet()
 
   url1 <- "bit.ly/DataDropPH"
   req1 <- httr::GET(url1)
@@ -64,7 +63,7 @@ LoadPhilippines <- function() {
           DateDied = vroom::col_date(format = "%Y-%m-%d"),
           DateRecover = vroom::col_date(format = "%Y-%m-%d"),
           DateOnset = vroom::col_date(format = "%Y-%m-%d")
-        ), 
+        ),
         show_col_types = FALSE
       )
       unlink(temp)
