### 2022-11-27 sjbeckett/localcovid19now#59 @sjbeckett
Thanks for reporting this @mbkumar. Will check this as we perform revisions.
