### 2022-11-28 bdd4e27
remove deprecated function makeMaps.R
