+++ R/MakeTable.R
@@ -9,7 +9,7 @@
 #' @export
 #' @seealso [estRisk()]
 #' @examples
-#' risk <- calcrisk(.001, 50)
+#' risk <- calcRisk(.001, 50)
 #'
 calcRisk <- function(p_I, g) {
   r <- 1 - (1 - p_I)**g
