### 2022-11-28 ab41a6c
added dontrun to countries which take a long time to load.
removed dontrun from estRisk, added examples to PerCapitaMap_leaflet and PerCapitaMap_tmap

add dontrun to Austria

add dontrun example tags and update tidy_Data

update dontrun tags

add dontrun to UK and Vietnam (elapsed >5 s)
