+++ R/LoadNetherlands.R
@@ -28,7 +28,9 @@ getDataND <- function(Mcode, netherlandsData) {
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
+#' \dontrun{
 #' Netherlands <- LoadNetherlands()
+#' }
 #' @seealso [LoadCountries()]
 #' @export
 LoadNetherlands <- function() {
