### 2022-11-28 4cf3d0c
removed dontrun from estRisk, added examples to PerCapitaMap_leaflet and PerCapitaMap_tmap
