### 2022-11-28 21fd128
Added LoadErrorWarning, and updated documentation
