+++ NAMESPACE
@@ -16,6 +16,7 @@ export(LoadCuba)
 export(LoadCzechia)
 export(LoadDenmark)
 export(LoadEcuador)
+export(LoadErrorWarning)
 export(LoadEurope)
 export(LoadFrance)
 export(LoadGermany)
