### 2022-11-28 4db7731
Added LoadErrorWarning, and updated documentation
