+++ man/EventMap_tmap.Rd
@@ -42,7 +42,6 @@ EventMap_tmap(Austria, 50, Austria$AB)
 \seealso{
 Other mapplots: 
 \code{\link{EventMap_leaflet}()},
-\code{\link{PerCapitaMap_leaflet}()},
 \code{\link{PerCapitaMap_tmap}()}
 }
 \concept{mapplots}
