### 2022-11-28 8d479af
fixed PerCapitaMap_leaflet documentation.
