+++ man/EventMap_leaflet.Rd
@@ -31,6 +31,7 @@ EventMap_leaflet(Austria, 50, Austria$AB)
 \seealso{
 Other mapplots: 
 \code{\link{EventMap_tmap}()},
+\code{\link{PerCapitaMap_leaflet}()},
 \code{\link{PerCapitaMap_tmap}()}
 }
 \concept{mapplots}
