### 2022-11-28 d718fa7
fixed PerCapitaMap_leaflet documentation.
