+++ R/LoadAustria.R
@@ -6,7 +6,9 @@
 #' Data is from the Federal Ministry for Social Affairs, Health, Care and Consumer Protection (BMSGPK) data on COVID-19 for Austria \url{https://www.data.gv.at/covid-19/}.
 #'
 #' @examples
+#' \dontrun{
 #' Austria <- LoadAustria()
+#' }
 #' @seealso [LoadCountries]
 #' @export
 LoadAustria <- function() {
