+++ man/LoadAustria.Rd
@@ -13,7 +13,9 @@ Reads in subnational data for Austria to calculate most recent estimate of per c
 Data is from the Federal Ministry for Social Affairs, Health, Care and Consumer Protection (BMSGPK) data on COVID-19 for Austria \url{https://www.data.gv.at/covid-19/}.
 }
 \examples{
+\dontrun{
 Austria <- LoadAustria()
+}
 }
 \seealso{
 \link{LoadCountries}
