+++ R/tidy_Data.R
@@ -3,6 +3,7 @@
 #' @description For an object read in with a LoadCountry function, sets the per capita active cases (pInf) to NA or removes them from the dataset based on input conditions. Can be used to remove data that is not recent, or data that produces low or negative estimates of per capita cases, or risk. Can also be used to remove data that is NA by country or by region.
 #'
 #' @param DATA Data created using a LoadCountry call.
+#' @param tidy If TRUE, then perform tidying according to other parameters. If FALSE, then do nothing.
 #' @param DaysOld  Set any pInf data more than this days old to NA.
 #' @param minimumpercapitaactivecases Set any pInf data less than this to NA.
 #' @param RiskEval Set pInf to NA when risk is below RiskEval$minimumRisk (%) using RiskEval$ascertainmentbias and a maximum group size, RiskEval$maximumN (Note: this setting overwrites minimumpercapitaactivecases).
@@ -13,18 +14,21 @@
 #'
 #' @examples
 #' \dontrun{
-#' Philippines <- LoadPhilippines()
-#' Philippines <- tidy_Data(Philippines)
+#' Europe1 <- LoadEurope() #tidy_Data is run within LoadX functions.
+#' Europe2 <- LoadEurope(tidy = FALSE)
+#' Europe2 <- tidy_Data(Europe2)
 #' }
 #' @export
-tidy_Data <- function(DATA, DaysOld = 30, minimumpercapitaactivecases = 0, RiskEval = NULL, dropNACountry = FALSE, dropNAall = FALSE) {
+tidy_Data <- function(DATA, tidy = TRUE, DaysOld = 30, minimumpercapitaactivecases = 0, RiskEval = NULL, dropNACountry = FALSE, dropNAall = FALSE) {
   # DATA is a dataset created from a LoadX function.
+  # Perform tidying only when tidy is TRUE.
   # Set any data older than DaysOld to Not Available.
   # Set any data where per capita active cases are below the minimumpercapitaactivecases to Not Available.
   # Set any data where Risk is below RiskEval$minimumRisk (%) using RiskEval$ascertainmentbias and a maximum group size, RiskEval$maximumN to Not Available. (Note: this setting overwrites minimumpercapitaactivecases).
   # remove rows for countries whose estimates all return NA
   # remove rows for any region whose estimates return NA
 
+  if (tidy == TRUE) {
 
   ## set NA conditions:
   # 1. by report date
@@ -61,5 +65,6 @@ tidy_Data <- function(DATA, DaysOld = 30, minimumpercapitaactivecases = 0, RiskE
     DATA <- DATA[-which(is.na(DATA$pInf)), ]
   }
 
+  }
   return(DATA)
 }
