+++ R/LoadCountries.R
@@ -8,7 +8,10 @@
 #' @export
 #'
 #' @examples
+#' USandMalaysia <- LoadCountries(countries = c("LoadUS","LoadMalaysia"))
+#' \dontrun{
 #' AllData <- LoadCountries()
+#' }
 LoadCountries <- function(countries = NULL, filepath = NULL, interactiveMode = interactive()) {
   # COMBINE DATASETS INTO SINGLE OBJECT
   NEWMAP <- c()
