+++ man/LoadSaudiArabia.Rd
@@ -16,7 +16,9 @@ Reads in subnational data for Saudi Arabia to calculate most recent estimate of 
 Data sourced from Ministry of Health, Covid19 Command and Control Center CCC, The National Health Emergency Operation Center NHEOC; and assembled by National Health Command and Control NHCC, Covid19 Data and Informatics Committee. \url{https://covid19.moh.gov.sa/}.
 }
 \examples{
+\dontrun{
 SaudiArabia <- LoadSaudiArabia()
+}
 }
 \seealso{
 \code{\link[=LoadCountries]{LoadCountries()}}
