+++ man/LoadCountries.Rd
@@ -24,5 +24,8 @@ A simple feature returning the date of most recent data (DateReport), a unique r
 LoadCountries
 }
 \examples{
+USandMalaysia <- LoadCountries(countries = c("LoadUS","LoadMalaysia"))
+\dontrun{
 AllData <- LoadCountries()
+}
 }
