### 2022-11-29 sjbeckett/localcovid19now#58 @welch16
I tried installing the package again after increasing the timeout and it installed correctly. I wonder if this is a common enough issue to mentioning it in the installation instructions

Thanks for your help
