### 2022-11-29 sjbeckett/localcovid19now#62 @welch16
Plans on adding unit tests?
Hi @sjbeckett,

I was wondering if there are any plans to add unit tests with `testthat` or an alternative package. I know that other organizations like bioconductor or ropensci recommend adding unit tests.

I know the amount of work required to add tests to a package of the scope as this one, but I feel it would be enriched in the end.

Thanks
