+++ R/LoadMexico.R
@@ -21,25 +21,25 @@ LoadMexico <- function() {
 
   geomMexico <- NULL
   utils::data("geomMexico", envir = environment())
+  geomMexico <- sf::st_as_sf(geomMexico)
 
   flag <- 0
   aa <- 1
+  MEX <- NULL
+  
   while (flag == 0) {
     DATE <- Sys.Date() - aa
     formDATE <- format(DATE, "%Y%m%d")
     STRING <- paste0("https://datos.covid-19.conacyt.mx/Downloads/Files/Casos_Diarios_Municipio_Confirmados_", formDATE)
 <<<<<<< HEAD
     MEX <- try(vroom::vroom(STRING, show_col_types = FALSE, progress = FALSE), silent = TRUE) # note older files are DELETED.
-=======
-    MEX <- try(vroom::vroom(STRING,show_col_types=FALSE)) # note older files are DELETED.
->>>>>>> 9cf4ba0 (fixed warning messages/ and removed other print outs to console.)
-    if (is.null(dim(MEX)) == FALSE) {
+    if (is.null(MEX) == FALSE) {
       flag <- 1
     } else {
       aa <- aa + 1
     }
     if (aa > 30) {
-      warning("no recent data")
+      stop("no recent data")
       flag <- 2
     }
   }
