### 2022-11-29 0eb7d56
add test files for Europe and Oman to test proposed new LoadX structure
