+++ R/LoadOmanTESTING.R
@@ -1,0 +1,66 @@
+#' LoadOmanTESTING
+#'
+#' @description Reads in subnational data for Oman to calculate most recent estimate of per capita active COVID-19 cases.
+#'
+#' @param tidy If TRUE, then perform tidying according to other parameters. If FALSE, then do nothing.
+#' @param DaysOld  Set any pInf data more than this days old to NA.
+#' @param minimumpercapitaactivecases Set any pInf data less than this to NA.
+#' @param RiskEval Set pInf to NA when risk is below RiskEval$minimumRisk (%) using RiskEval$ascertainmentbias and a maximum group size, RiskEval$maximumN (Note: this setting overwrites minimumpercapitaactivecases).
+#' @param dropNACountry If TRUE, remove rows for countries whose pInf estimates all return NA.
+#' @param dropNAall If TRUE, remove rows for any region whose pInf estimates all return NA
+#'
+#' @note
+#' Ministry of Health for Oman, collated by Safeture for the Humanitarian Data Exchange: \url{https://data.humdata.org/dataset/oman-coronavirus-covid-19-subnational-cases}.
+#'
+#' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
+#'
+#' @examples
+#' Oman2 <- LoadOmanTESTING()
+#' @seealso [LoadCountries()]
+#' @export
+LoadOmanTESTING <- function(tidy = TRUE, DaysOld = 30, minimumpercapitaactivecases = 0, RiskEval = NULL, dropNACountry = FALSE, dropNAall = FALSE) {
+  
+  tryCatch({
+  # Ministry of Health for Oman, collated by Safeture for the Humanitarian Data Exchange: https://data.humdata.org/dataset/oman-coronavirus-covid-19-subnational-cases
+  
+  utils::data("geomOman", envir = environment())
+  
+  CaseData <- utils::read.csv("https://www.dropbox.com/s/ylop7xswywi147c/cases_oman.csv?dl=1")
+  Governorates <- sort(unique(CaseData$name))
+  CaseDiff <- c()
+  DateReport <- c()
+  
+  for (aa in 1:length(Governorates)) {
+    subsetdata <- CaseData[which(CaseData$name == Governorates[aa]), ]
+    CaseDiff[aa] <- (10 / 14) * sum(utils::tail(subsetdata$cases, 14))
+    DateReport[aa] <- utils::tail(subsetdata$date, 1)
+  }
+  omandt <- data.frame(DateReport, Name = Governorates, Difference = CaseDiff)
+  
+  # population
+  pop <- c()
+  pop$Name <- c("Ad Dakhiliyah", "Al Buraymi", "Al Wusta", "Ad Dhahirah", "Al Batinah South", "Ash Sharqiyah South", "Muscat", "Musandam", "Al Batinah North", "Ash Sharqiyah North", "Dhofar")
+  pop$Population <- c(478501, 121802, 52344, 213043, 465550, 315445, 1302440, 49062, 784681, 271822, 416458) # 2020 census - https://www.citypopulation.de/en/oman/cities/
+  pop <- data.frame(pop)
+  
+  omandf <- dplyr::inner_join(omandt, pop, by = c("Name" = "Name"))
+  
+  # geometry
+  # https://data.humdata.org/dataset/cod-ab-omn?
+  # data("geomOman")
+  geomOman$micro_name[5] <- "Ad Dhahirah"
[…]
