+++ NAMESPACE
@@ -18,6 +18,8 @@ export(LoadDenmark)
 export(LoadEcuador)
 export(LoadErrorWarning)
 export(LoadEurope)
+export(LoadEuropeTESTING)
+export(LoadEuropeTESTINGBROKEN)
 export(LoadFrance)
 export(LoadGermany)
 export(LoadGhana)
@@ -36,6 +38,7 @@ export(LoadNewZealand)
 export(LoadNigeria)
 export(LoadNorway)
 export(LoadOman)
+export(LoadOmanTESTING)
 export(LoadPakistan)
 export(LoadPeru)
 export(LoadPhilippines)
