### 2022-12-04 6e6e17e
provide type checking for tidy_Data
