### 2022-12-04 b678179
provide type checking for tidy_Data
added missing instances of 'geomA <- sf::st_as_sf(geomA)'

add dontrun to UK and Vietnam (elapsed >5 s)

small typechecking to estRisk and calcRisk

Updates to include a LoadDataHelper function to help modularize code. Tidy type checking for tidy_Data. Update default behaviour for tidy_Data to drop country if all NA values. Update documentation.

catch errors in tidy_Data, update assertions for RiskEval, make sure tidy_Data options are passed along in LoadDataHelper

assertOptions example too long and add Denmark to dontrun

Small fixes to LoadX files to red
[…]
