+++ R/LoadMexico.R
@@ -1,6 +1,6 @@
 #' LoadMexico
 #'
-#' @description Reads in subnational data for Mexico to calculate most recent estimate of per capita active COVID-19 cases.
+#' @description Reads in subnational data for Mexico to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
 #'
 #' @note
 #' COVID-19 data is sourced from the Covid-19 México hub page: \url{https://datos.covid-19.conacyt.mx/}.
@@ -11,7 +11,7 @@
 #' \dontrun{
 #' Mexico <- LoadMexico()
 #' }
-#' @seealso [LoadCountries()]
+#' @seealso [LoadData()]
 #' @export
 LoadMexico <- function() {
   # COVID-19 Covid-19 México hub page: https://datos.covid-19.conacyt.mx/
@@ -21,21 +21,24 @@ LoadMexico <- function() {
 
   geomMexico <- NULL
   utils::data("geomMexico", envir = environment())
+  geomMexico <- sf::st_as_sf(geomMexico)
 
   flag <- 0
   aa <- 1
+  MEX <- NULL
+  
   while (flag == 0) {
     DATE <- Sys.Date() - aa
     formDATE <- format(DATE, "%Y%m%d")
     STRING <- paste0("https://datos.covid-19.conacyt.mx/Downloads/Files/Casos_Diarios_Municipio_Confirmados_", formDATE)
     MEX <- try(vroom::vroom(STRING, show_col_types = FALSE, progress = FALSE), silent = TRUE) # note older files are DELETED.
-    if (is.null(dim(MEX)) == FALSE) {
+    if (is.null(MEX) == FALSE) {
       flag <- 1
     } else {
       aa <- aa + 1
     }
     if (aa > 30) {
-      warning("no recent data")
+      stop("no recent data")
       flag <- 2
     }
   }
