+++ R/LoadVietnam.R
@@ -1,6 +1,6 @@
 #' LoadVietnam
 #'
-#' @description Reads in subnational data for Vietnam to calculate most recent estimate of per capita active COVID-19 cases.
+#' @description Reads in subnational data for Vietnam to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
 #'
 #' @note
 #' Official data on the COVID-19 epidemic in Vietnam collected by the Vietnam Ministry of Health & Ministry of Information and Communications. Operated by National Center for Technology for COVID-19 Prevention and Control. Developed by VN National Cyber Security Center. \url{https://covid19.ncsc.gov.vn/dulieu}.
@@ -8,8 +8,10 @@
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
+#' \dontrun{
 #' Vietnam <- LoadVietnam()
-#' @seealso [LoadCountries()]
+#' }
+#' @seealso [LoadData()]
 #' @export
 LoadVietnam <- function() {
   # Official data on the COVID-19 epidemic in Vietnam collected by the Vietnam Ministry of Health & Ministry of Information and Communications. Operated by National Center for Technology for COVID-19 Prevention and Control. Developed by VN National Cyber Security Center. https://covid19.ncsc.gov.vn/dulieu
@@ -29,6 +31,7 @@ LoadVietnam <- function() {
   misc_vietnam <- geomVietnam <- NULL
   utils::data("misc_vietnam", envir = environment())
   utils::data("geomVietnam", envir = environment())
+  geomVietnam <- sf::st_as_sf(geomVietnam)
 
   # ProvinceInfo <- misc_vietnam
   # Geometry

[…]
