### 2022-12-04 6955f37
added missing instances of 'geomA <- sf::st_as_sf(geomA)'
