+++ R/LoadOmanTESTING.R
@@ -24,6 +24,7 @@ LoadOmanTESTING <- function(tidy = TRUE, DaysOld = 30, minimumpercapitaactivecas
   # Ministry of Health for Oman, collated by Safeture for the Humanitarian Data Exchange: https://data.humdata.org/dataset/oman-coronavirus-covid-19-subnational-cases
   
   utils::data("geomOman", envir = environment())
+  geomOman <- sf::st_as_sf(geomOman)
   
   CaseData <- utils::read.csv("https://www.dropbox.com/s/ylop7xswywi147c/cases_oman.csv?dl=1")
   Governorates <- sort(unique(CaseData$name))
