### 2022-12-04 b036dc7
add dontrun to UK and Vietnam (elapsed >5 s)
