+++ R/LoadUK.R
@@ -43,7 +43,9 @@ dataQueryUK <- function(date) {
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
+#' \dontrun{
 #' UK <- LoadUK()
+#' }
 #' @seealso [LoadCountries()]
 #' @export
 LoadUK <- function() {
