### 2022-12-04 90b4ddb
small typechecking to estRisk and calcRisk
