### 2022-12-05 ced0ed5
Updates to include a LoadDataHelper function to help modularize code. Tidy type checking for tidy_Data. Update default behaviour for tidy_Data to drop country if all NA values. Update documentation.
