+++ NAMESPACE
@@ -14,6 +14,7 @@ export(LoadChina)
 export(LoadCountries)
 export(LoadCuba)
 export(LoadCzechia)
+export(LoadDataHelper)
 export(LoadDenmark)
 export(LoadEcuador)
 export(LoadErrorWarning)
@@ -58,6 +59,7 @@ export(LoadZimbabwe)
 export(PerCapitaMap_leaflet)
 export(PerCapitaMap_tmap)
 export(addNewGeoms)
+export(assertOptions__tidy_Data)
 export(calcRisk)
 export(create_c19r_data)
 export(estRisk)

[…]
