### 2022-12-06 065c79d
catch errors in tidy_Data, update assertions for RiskEval, make sure tidy_Data options are passed along in LoadDataHelper
