### 2022-12-06 ed0a5e5
assertOptions example too long and add Denmark to dontrun
