### 2022-12-07 4bf000a
Add LoadData as a replacement to LoadCountries; and base functionality around this function rather than individual source files i.e. LoadData("LoadItaly") rather than LoadItaly().
cast text in LoadErrorWarning to message

Update README, NAMESPACE and examples

remove LoadDataHelper

remove testing files
