+++ man/LoadDataHelper.Rd
@@ -1,47 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/LoadDataHelper.R
-\name{LoadDataHelper}
-\alias{LoadDataHelper}
-\title{LoadDataHelper}
-\usage{
-LoadDataHelper(
-  functionName,
-  tidy = TRUE,
-  DaysOld = 30,
-  minimumpercapitaactivecases = 0,
-  RiskEval = NULL,
-  dropNACountry = TRUE,
-  dropNAall = FALSE
-)
-}
-\arguments{
-\item{functionName}{Name of function representing that for loading data.}
-
-\item{tidy}{If TRUE, then perform tidying according to other parameters. If FALSE, then do nothing. (passed to tidy_Data).}
-
-\item{DaysOld}{Set any pInf data more than this days old to NA.(passed to tidy_Data).}
-
-\item{minimumpercapitaactivecases}{Set any pInf data less than this to NA.(passed to tidy_Data).}
-
-\item{RiskEval}{Set pInf to NA when risk is below RiskEval$minimumRisk (\%) using RiskEval$ascertainmentbias and a maximum group size, RiskEval$maximumN (Note: this setting overwrites minimumpercapitaactivecases). (passed to tidy_Data).}
-
-\item{dropNACountry}{If TRUE, remove rows for countries whose pInf estimates all return NA.(passed to tidy_Data).}
-
-\item{dropNAall}{If TRUE, remove rows for any region whose pInf estimates all return NA. (passed to tidy_Data).}
-}
-\value{
-nothing.
-}
-\description{
-General framework for loading and tidying data.
-}
-\examples{
-LoadDataHelper("LoadUS")
-
-RE = c()
-RE$minimumRisk = 5
-RE$ascertainmentbias = 5
-RE$maximumN = 100
-LoadDataHelper("LoadUS", RiskEval=RE)
-}
-\keyword{internal}
