+++ NAMESPACE
@@ -14,13 +14,11 @@ export(LoadChina)
 export(LoadCountries)
 export(LoadCuba)
 export(LoadCzechia)
-export(LoadDataHelper)
+export(LoadData)
 export(LoadDenmark)
 export(LoadEcuador)
 export(LoadErrorWarning)
 export(LoadEurope)
-export(LoadEuropeTESTING)
-export(LoadEuropeTESTINGBROKEN)
 export(LoadFrance)
 export(LoadGermany)
 export(LoadGhana)
@@ -39,7 +37,6 @@ export(LoadNewZealand)
 export(LoadNigeria)
 export(LoadNorway)
 export(LoadOman)
-export(LoadOmanTESTING)
 export(LoadPakistan)
 export(LoadPeru)
 export(LoadPhilippines)

[…]
