### 2022-12-07 294fe8e
Add LoadData as a replacement to LoadCountries; and base functionality around this function rather than individual source files i.e. LoadData("LoadItaly") rather than LoadItaly().
