+++ R/LoadData.R
@@ -1,0 +1,103 @@
+#' LoadData
+#'
+#' @description General framework for loading and tidying data.
+#'
+#' @param functionNames Name(s) of function representing that for loading data - options are in countrylist. Use NULL to attempt to download all available datasets.
+#' @param filepath Optionally, provide a filepath to save an error csv file to.
+#' @param interactiveMode Set whether the session is being run interactively. If not and no googledrive oauth token is found, avoid data requiring googledrive auth token.
+#' @param tidy If TRUE, then perform tidying according to other parameters. If FALSE, then do nothing. (passed to tidy_Data).
+#' @param DaysOld  Set any pInf data more than this days old to NA.(passed to tidy_Data).
+#' @param minimumpercapitaactivecases Set any pInf data less than this to NA.(passed to tidy_Data).
+#' @param RiskEval Set pInf to NA when risk is below RiskEval$minimumRisk (%) using RiskEval$ascertainmentbias and a maximum group size, RiskEval$maximumN (Note: this setting overwrites minimumpercapitaactivecases). (passed to tidy_Data).
+#' @param dropNACountry If TRUE, remove rows for countries whose pInf estimates all return NA.(passed to tidy_Data).
+#' @param dropNAall If TRUE, remove rows for any region whose pInf estimates all return NA. (passed to tidy_Data).
+#' @param custom If TRUE, allows for use with functions not listed in object countrylist (experimental usage).
+#' @param verbose If TRUE, reports on loading progress and returns warnings/errors to console.
+#' @keywords internal
+#' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
+#'
+#' @examples
+#' LoadData("LoadUS")
+#' LoadData(c("LoadUS","LoadMalaysia"))
+#' \dontrun{
+#' LoadData()
+#' }
+#' @export
+LoadData <- function(functionNames = NULL, filepath = NULL, interactiveMode = interactive(), tidy = TRUE, DaysOld = 30, minimumpercapitaactivecases = 0, RiskEval = NULL, dropNACountry = TRUE, dropNAall = FALSE, custom = FALSE, verbose = TRUE){
+  
+  #check inputs
+  assertOptions__tidy_Data(tidy,DaysOld,minimumpercapitaactivecases,RiskEval,dropNACountry,dropNAall)
+  stopifnot("`interactiveMode` must be a logical." = is.logical(interactiveMode))
+  stopifnot("`custom` must be a logical." = is.logical(custom))
+  stopifnot("`verbose` must be a logical." = is.logical(verbose))
+  
+  
+  # COMBINE DATASETS INTO SINGLE OBJECT
+  NEWMAP <- c()
+  
+  # assign country list if not provided
+  if (is.null(functionNames)) {
+    utils::data(countrylist, envir = environment())
+    functionNames <- countrylist
+  }
+  
+  # check Load options that are input match allowable options.
+  if (custom == FALSE){
+    stopifnot("`functionNames` must match options in object countrylist, or be set to NULL." = all(functionNames %in% countrylist))
+  }
+  
+  # check whether googledrive credentials are available if in non-interactive mode.
+  if (interactiveMode == FALSE) {
+    TOKEN <- googledrive::drive_has_token()
+    if (TOKEN == FALSE) {
+      warning("No googledrive token found. Will avoid loading data requiring token. For more about providing authorization, see: https://gargle.r-lib.org/articles/non-interactive-auth.html")
[…]
