### 2022-12-07 b2ad8e9
Update README, NAMESPACE and examples
