+++ NAMESPACE
@@ -14,6 +14,7 @@ export(LoadChina)
 export(LoadCountries)
 export(LoadCuba)
 export(LoadCzechia)
+export(LoadData)
 export(LoadDataHelper)
 export(LoadDenmark)
 export(LoadEcuador)
