### 2022-12-07 293f004
remove LoadDataHelper
