+++ NAMESPACE
@@ -15,7 +15,6 @@ export(LoadCountries)
 export(LoadCuba)
 export(LoadCzechia)
 export(LoadData)
-export(LoadDataHelper)
 export(LoadDenmark)
 export(LoadEcuador)
 export(LoadErrorWarning)
