+++ man/LoadEuropeTESTINGBROKEN.Rd
@@ -1,45 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/LoadEuropeTESTINGBROKEN.R
-\name{LoadEuropeTESTINGBROKEN}
-\alias{LoadEuropeTESTINGBROKEN}
-\title{LoadEuropeTESTINGBROKEN}
-\usage{
-LoadEuropeTESTINGBROKEN(
-  tidy = TRUE,
-  DaysOld = 30,
-  minimumpercapitaactivecases = 0,
-  RiskEval = NULL,
-  dropNACountry = FALSE,
-  dropNAall = FALSE
-)
-}
-\arguments{
-\item{tidy}{If TRUE, then perform tidying according to other parameters. If FALSE, then do nothing.}
-
-\item{DaysOld}{Set any pInf data more than this days old to NA.}
-
-\item{minimumpercapitaactivecases}{Set any pInf data less than this to NA.}
-
-\item{RiskEval}{Set pInf to NA when risk is below RiskEval$minimumRisk (\%) using RiskEval$ascertainmentbias and a maximum group size, RiskEval$maximumN (Note: this setting overwrites minimumpercapitaactivecases).}
-
-\item{dropNACountry}{If TRUE, remove rows for countries whose pInf estimates all return NA.}
-
-\item{dropNAall}{If TRUE, remove rows for any region whose pInf estimates all return NA}
-}
-\value{
-A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
-}
-\description{
-Reads in subnational data for Europe to calculate most recent estimate of per capita active COVID-19 cases.
-}
-\note{
-Data aggregated from local health resources in the WHO European Region COVID19 Subnational Explorer \url{https://experience.arcgis.com/experience/3a056fc8839d47969ef59949e9984a71}.
-}
-\examples{
-\dontrun{
-Europe3 <- LoadEuropeTESTINGBROKEN()
-}
-}
-\seealso{
-\code{\link[=LoadCountries]{LoadCountries()}}
-}
