+++ NAMESPACE
@@ -19,8 +19,6 @@ export(LoadDenmark)
 export(LoadEcuador)
 export(LoadErrorWarning)
 export(LoadEurope)
-export(LoadEuropeTESTING)
-export(LoadEuropeTESTINGBROKEN)
 export(LoadFrance)
 export(LoadGermany)
 export(LoadGhana)
@@ -39,7 +37,6 @@ export(LoadNewZealand)
 export(LoadNigeria)
 export(LoadNorway)
 export(LoadOman)
-export(LoadOmanTESTING)
 export(LoadPakistan)
 export(LoadPeru)
 export(LoadPhilippines)
