### 2022-12-07 sjbeckett/localcovid19now#63 @jeanpaulrsoucy
Canadian data updates
Hello,

I'm the maintainer of [`CovidTimelineCanada`](https://github.com/ccodwg/CovidTimelineCanada), which you appear to use to source Canadian data in your project.

I just wanted to let you know that our health regions data file `geo/health_regions.csv` has been renamed `geo/hr.csv`. The same is true of the geoJSON map file. You refer the old files in two places in the code:

https://github.com/sjbeckett/localcovid19now/blob/e93ecf10b2b6a86c5139475c24648f9ad4eb39a4/R/LoadCanada.R#L32

https://github.com/sjbeckett/localcovid19now/blob/e93ecf10b2b6a86c5139475c24648f9ad4eb39a4/tools/processed/processCanada.R#L2

The new links are, respectively:

https://raw.githubusercontent.com/ccodwg/CovidTimelineCanada/main/geo/hr.csv

https://raw.githubusercontent.com/ccodwg/CovidTimelineCanada/main/geo/hr.geojson

The old files will exist as a duplicate for the foreseeable future, but I figured it was best to be proactive and open up an issue for this. I am renaming the file as part of an effort to make the file names more consistent (see ccodwg/CovidTimelineCanada#62).

I would have made the change myself as a pull request, but I noticed a few potential issues. First, `hr.csv` (and `health_regions.csv` recently had a "pruid" column added, which gives numerical IDs of provinces (similar to the numerical IDs of health regions that exist now). This could cause an issue with your join here:

https://github.com/sjbeckett/localcovid19now/blob/e93ecf10b2b6a86c5139475c24648f9ad4eb39a4/R/LoadCanada.R#L61

...if `CANDT` has a column called "pruid", but I don't think it does.

Earlier in the same script, I noticed:

https://github.com/sjbeckett/localcovid19now/blob/e93ecf10b2b6a86c5139475c24648f9ad4eb39a4/R/LoadCanada.R#L35

The `CovidTimelineCanada` dataset has always had data in YYYY-MM-DD format, but the earlier [`Covid19Canada`](https://github.com/ccodwg/Covid19Canada) dataset used `DD-MM-YYYY`, so I'm not sure if this may need to be updated as well.

Please let me know if you have any questions or issues.
