### 2022-12-07 sjbeckett/localcovid19now#63 @sjbeckett
Many thanks for reaching out @jeanpaulrsoucy. Yes, we first started interacting with the earlier `Covid19Canada`, but looks like we haven't fully integrated with the new package. We are currently revising some of the ways in which our package is organised, so I will make updates to the relevant files within our current draft PR. Thank you for pointing out potential conflicts in our code base; and thank you for your work with both `Covid19Canada` and `CovidTimeLineCanada'!
