### 2022-12-07 fa300a3
update Canada loading to follow updates to CovidTimelineCanada
