### 2022-12-08 0fa784e
add internal to functions regarding updating geometry - more for development purposes
