+++ man/updateGlobal.Rd
@@ -14,3 +14,4 @@ Update Global Geometries
 updateGlobal()
 }
 }
+\keyword{internal}
