### 2022-12-08 5650341
add internal to functions regarding updating geometry - more for development purposes
