+++ man/moveFiles.Rd
@@ -16,3 +16,4 @@ Move Processed Files
 file.create(here::here("tools", "toProcess", "testFile.R"))
 moveFiles("testFile.R")
 }
+\keyword{internal}
