+++ R/LoadUK.R
@@ -35,7 +35,7 @@ dataQueryUK <- function(date) {
 
 #' LoadUK
 #'
-#' @description Reads in subnational data for the United Kingdom to calculate most recent estimate of per capita active COVID-19 cases.
+#' @description Reads in subnational data for the United Kingdom to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
 #'
 #' @note
 #' The COVID-19 data is from the UK API from Public Health England and NHSX: \url{https://coronavirus.data.gov.uk}.
@@ -46,7 +46,7 @@ dataQueryUK <- function(date) {
 #' \dontrun{
 #' UK <- LoadUK()
 #' }
-#' @seealso [LoadCountries()]
+#' @seealso [LoadData()]
 #' @export
 LoadUK <- function() {
   # The COVID-19 data is from the UK API from Public Health England and NHSX: https://coronavirus.data.gov.uk
