+++ R/LoadAustria.R
@@ -1,6 +1,6 @@
 #' LoadAustria
 #'
-#' @description Reads in subnational data for Austria to calculate most recent estimate of per capita active COVID-19 cases.
+#' @description Reads in subnational data for Austria to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
 #'
 #' @note
 #' Data is from the Federal Ministry for Social Affairs, Health, Care and Consumer Protection (BMSGPK) data on COVID-19 for Austria \url{https://www.data.gv.at/covid-19/}.
@@ -9,7 +9,7 @@
 #' \dontrun{
 #' Austria <- LoadAustria()
 #' }
-#' @seealso [LoadCountries]
+#' @seealso [LoadData]
 #' @export
 LoadAustria <- function() {
   # Load in geometry
