+++ R/LoadFrance.R
@@ -1,6 +1,6 @@
 #' LoadFrance
 #'
-#' @description Reads in subnational data for France to calculate most recent estimate of per capita active COVID-19 cases.
+#' @description Reads in subnational data for France to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
 #'
 #' @note
 #' Santé publique France COVID-19 data for France :  \url{https://www.data.gouv.fr/fr/datasets/donnees-relatives-aux-resultats-des-tests-virologiques-covid-19/} // \url{https://www.data.gouv.fr/fr/datasets/donnees-de-laboratoires-pour-le-depistage-a-compter-du-18-05-2022-si-dep/}.
@@ -10,7 +10,7 @@
 #'
 #' @examples
 #' France <- LoadFrance()
-#' @seealso [LoadCountries()]
+#' @seealso [LoadData()]
 #' @export
 LoadFrance <- function() {
   # Santé publique France COVID-19 data for France :  https://www.data.gouv.fr/fr/datasets/donnees-relatives-aux-resultats-des-tests-virologiques-covid-19/ // https://www.data.gouv.fr/fr/datasets/donnees-de-laboratoires-pour-le-depistage-a-compter-du-18-05-2022-si-dep/
