### 2022-12-08 18a222b
updates Load functions to point to LoadData
update readme install instructions

update examples in mapping and tables to use LoadData()

removed LoadCountries

update Namespace
