+++ R/LoadAustralia.R
@@ -1,13 +1,13 @@
 #' LoadAustralia
 #'
-#' @description Reads in subnational data for Australia to calculate most recent estimate of per capita active COVID-19 cases.
+#' @description Reads in subnational data for Australia to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
 #'
 #' @note
 #' Data was obtained from \url{https://github.com/M3IT/COVID-19_Data} and aggregated by www.covid19data.com.au from local health resources.
 #'
 #' @examples
 #' Australia <- LoadAustralia()
-#' @seealso [LoadCountries]
+#' @seealso [LoadData()]
 #' @export
 LoadAustralia <- function() {
   # Load in geomtry and population data
