+++ R/LoadSwitzerlandLiechtenstein.R
@@ -1,6 +1,6 @@
 #' LoadSwitzerlandLiechtenstein
 #'
-#' @description Reads in subnational data for Switzerland and also data for Liechtenstein to calculate most recent estimate of per capita active COVID-19 cases.
+#' @description Reads in subnational data for Switzerland and also data for Liechtenstein to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
 #'
 #' @note
 #' Data obtained via the Federal Office of Public Health FOPH \url{https://www.covid19.admin.ch/en/overview}.
@@ -9,7 +9,7 @@
 #'
 #' @examples
 #' SwitzerlandLiechtenstein <- LoadSwitzerlandLiechtenstein()
-#' @seealso [LoadCountries()]
+#' @seealso [LoadData()]
 #' @export
 LoadSwitzerlandLiechtenstein <- function() {
   # Federal Office of Public Health FOPH https://www.covid19.admin.ch/en/overview
