### 2022-12-08 fdc2b11
update readme install instructions
