### 2022-12-08 7002f30
update readme install instructions
update examples in mapping and tables to use LoadData()

removed LoadCountries

update Namespace

added alternate list of names to match to existing function calls in LoadData.

update input checking in LoadData

switch argument ordering in LoadData, such that 2nd argument must be logical.

updated examples from a load on 9Dec2022
