+++ R/LoadData.R
@@ -25,32 +25,33 @@
 #' }
 #' @seealso [tidy_Data()]
 #' @export
-LoadData <- function(functionNames = NULL, filepath = NULL, interactiveMode = interactive(), tidy = TRUE, DaysOld = 30, minimumpercapitaactivecases = 0, RiskEval = NULL, dropNACountry = TRUE, dropNAall = FALSE, custom = FALSE, verbose = TRUE){
+LoadData <- function(functionNames = NULL, custom = FALSE, filepath = NULL, interactiveMode = interactive(), tidy = TRUE, DaysOld = 30, minimumpercapitaactivecases = 0, RiskEval = NULL, dropNACountry = TRUE, dropNAall = FALSE, verbose = TRUE){
   
   #check inputs
+  stopifnot("`custom` must be a logical." = is.logical(custom))
+  stopifnot("`filepath` must be a character string, or be set to null." = is.character(filepath) || is.null(filepath))
   assertOptions__tidy_Data(tidy,DaysOld,minimumpercapitaactivecases,RiskEval,dropNACountry,dropNAall)
   stopifnot("`interactiveMode` must be a logical." = is.logical(interactiveMode))
-  stopifnot("`custom` must be a logical." = is.logical(custom))
   stopifnot("`verbose` must be a logical." = is.logical(verbose))
-  
+
   
   # COMBINE DATASETS INTO SINGLE OBJECT
   NEWMAP <- c()
   
   utils::data(countrylist, envir = environment())
-
+  
   # assign country list if not provided
   if (is.null(functionNames)) {
-    utils::data(countrylist, envir = environment())
     functionNames <- countrylist
   }
   
   indcheck <- which(!(functionNames %in% countrylist))
   if(length(indcheck)>0){ #if not on main list, check on the alt_countrylist
+    utils::data(alt_countrylist, envir = environment())
       for(aa in indcheck){
-        thisC <- which(localcovid19now::alt_countrylist$inputCase == functionNames[aa])
+        thisC <- which(alt_countrylist$case == functionNames[aa])
         if(length(thisC)>0){
-          functionNames[aa] <- localcovid19now::alt_countrylist$loadingCall[thisC]
+          functionNames[aa] <- alt_countrylist$call[thisC]
         }
       }
   }
