+++ man/EventMap_tmap.Rd
@@ -34,7 +34,7 @@ Creates a tmap map displaying active cases per 'people' for the input data.
 }
 \examples{
 \dontrun{
-Austria <- LoadAustria()
+Austria <- LoadData("LoadAustria")
 Austria$AB <- 4
 EventMap_tmap(Austria, 50, Austria$AB)
 }
