### 2022-12-08 a792203
update examples in mapping and tables to use LoadData()
