+++ NAMESPACE
@@ -11,7 +11,6 @@ export(LoadBrazil)
 export(LoadCanada)
 export(LoadChile)
 export(LoadChina)
-export(LoadCountries)
 export(LoadCuba)
 export(LoadCzechia)
 export(LoadData)
