### 2022-12-09 59f555c
updated Canada, such that NT and NU aren't dropped from final dataset.
added alternate list of names to match to existing function calls in LoadData.
