+++ R/data.R
@@ -1,3 +1,6 @@
+#' alt_countrylist
+"alt_countrylist"
+
 #' countrylist
 "countrylist"
 
