+++ man/alt_countrylist.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{alt_countrylist}
+\alias{alt_countrylist}
+\title{alt_countrylist}
+\format{
+An object of class \code{data.frame} with 192 rows and 2 columns.
+}
+\usage{
+alt_countrylist
+}
+\description{
+alt_countrylist
+}
+\keyword{datasets}
