### 2022-12-09 85b00ab
update input checking in LoadData
switch argument ordering in LoadData, such that 2nd argument must be logical.

updated examples from a load on 9Dec2022

removed rgdal from Imports

update dontrun based on elapsed >5s

remove unused data

updated data and how alt_countrylist is called in LoadData
