+++ man/LoadChile.Rd
@@ -17,7 +17,9 @@ Data obtained from COVID-19 Data Hub \url{https://covid19datahub.io}
 sourced from Ministerio de Ciencia, Tecnología, Conocimiento e Innovación: \url{https://www.minciencia.gob.cl/COVID19/}.
 }
 \examples{
+\dontrun{
 Chile <- LoadChile()
+}
 }
 \references{
 Guidotti et al., (2020). COVID-19 Data Hub. Journal of Open Source Software, 5(51), 2376, \url{https://doi.org/10.21105/joss.02376}.
