### 2022-12-09 fa1dbfa
update input checking in LoadData
