### 2022-12-09 43d583d
switch argument ordering in LoadData, such that 2nd argument must be logical.
