### 2022-12-09 3b21b51
updated examples from a load on 9Dec2022
