### 2022-12-09 326458f
update dontrun based on elapsed >5s
