+++ R/LoadChile.R
@@ -12,7 +12,9 @@
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
+#' \dontrun{
 #' Chile <- LoadChile()
+#' }
 #' @seealso [LoadData()]
 #' @export
 LoadChile <- function() {
