+++ R/LoadSouthKorea.R
@@ -8,7 +8,9 @@
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
+#' \dontrun{
 #' SouthKorea <- LoadSouthKorea()
+#' }
 #' @seealso [LoadData()]
 #' @export
 LoadSouthKorea <- function() {
