+++ man/India.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\docType{data}
-\name{India}
-\alias{India}
-\title{India}
-\format{
-An object of class \code{character} of length 36.
-}
-\usage{
-India
-}
-\description{
-India
-}
-\keyword{datasets}
