+++ man/Haiti.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\docType{data}
-\name{Haiti}
-\alias{Haiti}
-\title{Haiti}
-\format{
-An object of class \code{numeric} of length 10.
-}
-\usage{
-Haiti
-}
-\description{
-Haiti
-}
-\keyword{datasets}
