### 2022-12-09 f5a5014
updated data and how alt_countrylist is called in LoadData
