+++ R/data.R
@@ -70,7 +70,6 @@
 #' geomHaiti
 "geomHaiti"
 
-
 #' geomIndia
 "geomIndia"
 
@@ -112,7 +111,6 @@
 
 #' geomPakistan
 "geomPakistan"
-
 
 #' geomPeru
 "geomPeru"
@@ -161,12 +159,6 @@
 
 #' geomZimbabwe
 "geomZimbabwe"
-
-#' Haiti
-"Haiti"
-
-#' India
-"India"
 
 #' m49
 "m49"
