### 2022-12-11 de1dbc1
condense geometry update functions to geomUpdateUtils.R; add description text to geometry update functions; add note to geomTemplate snippet to save in toProcess folder that might need to be created; renamed "moveFiles" to "moveProcessedFiles", "remSurplus" to "remGeoSurplus"
