+++ man/updateGlobal.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/updateGlobal.R
+% Please edit documentation in R/geomUpdateUtils.R
 \name{updateGlobal}
 \alias{updateGlobal}
 \title{Update Global Geometries}
@@ -7,7 +7,7 @@
 updateGlobal()
 }
 \description{
-Update Global Geometries
+Function to integrate in new geometries and update the geometries across the package
 }
 \examples{
 \dontrun{

[…]
