### 2022-12-11 dd717f0
condense geometry update functions to geomUpdateUtils.R; add description text to geometry update functions; add note to geomTemplate snippet to save in toProcess folder that might need to be created; renamed "moveFiles" to "moveProcessedFiles", "remSurplus" to "remGeoSurplus"
make update geometry functions internal; fix issue that would show no errors when files were not copied (instead of overwriting) but would be deleted anyway; change messages to use `message` instead of `cat`

fix compiling error (added dontrun to internal functions)

move europe raw files to inst/extdata; add data-raw to 
[…]
