+++ R/updateGlobal.R
@@ -1,153 +1,0 @@
-#' Update Global Geometries
-#'
-#' @export
-#'
-#' @keywords internal
-#' @examples
-#' \dontrun{
-#' updateGlobal()
-#' }
-updateGlobal <- function() {
-  filename <- geoid <- m49code <- iso3 <- country_name <- macro_code <- macro_name <- micro_code <- micro_name <- NULL
-
-  # Requires rmapshaper for the ms_simplify function - simplifies polygons.
-  rlang::check_installed(c("rmapshaper", "here"), reason = "to use `updateGlobal()`")
-
-  # Open the most recent WorldPreSimplified geojson. Manually modify if this is not the most up-to-date file.
-  ## Because zipfile named with seconds post epoch, `max(dir("updateGeometry/WorldPreSimplified"))` gives most recent version. Using ctime property potentially not reliable bc file creation with git.
-  zipfiles <- list.files(here::here("tools/WorldPreSimplified"))
-
-  temp <- utils::unzip(zipfile = here::here("tools/WorldPreSimplified", max(zipfiles)), exdir = tempdir())
-
-  geomWorld <- sf::st_read(temp)
-  unlink(temp)
-
-  ### Instructions
-  ## Add original geomCountry.geojson to countries/data/orig_geom
-  ## Create new script "updateGeometry/toProcess/processCounty.R" that matches processTemplate's format.
-  ## Run updateGlobal.R
-
-  # source("updateGeometry/addNewGeom.R")
-  newWorld <- addNewGeoms()
-
-  # If new geometries are added in newWorld, remove the old ones from geomWorld and add the new ones
-  if (!is.null(newWorld)) { # If no new geometries are added, newWorld will be NULL
-    updatedFiles <- unique(newWorld$filename)
-    geomWorld <- geomWorld %>%
-      dplyr::filter(!filename %in% updatedFiles) %>%
-      dplyr::bind_rows(
-        newWorld
-      )
-
-    # Remove geomSmallCountries and geomEurope entries that are supplied by other data sets
-    # source("updateGeometry/remSurplus.R")
-    geomWorld <- remSurplus(geomWorld, c("geomEurope", "geomSmallCountries"))
-
-    # Write a new geomWorld_presimp that includes the updated geometry
-    sf::st_write(geomWorld, "geomWorld_presimp.geojson", delete_dsn = T)
-
-    nowtime <- round(difftime(lubridate::now(tzone = "UTC"), lubridate::ymd_hms("1970-01-01 00:00:00"), tz = "UTC", units = "mins"))
-    # the presimplified world geometry is too big for github, so put it in a zip file and remove the geojson
-    zip(zipfile = paste0("tools/WorldPreSimplified/WorldPreSimp", nowtime, ".zip"), "geomWorld_presimp.geojson")
-    file.remove("geomWorld_presimp.geojson")
-    rm(newWorld)
-  } else {
-    updatedFiles <- NULL
-  }
-
-  # Simplify the world
-  geomGlobal <- rmapshaper::ms_simplify(geomWorld, keep = 0.05, explode = T, keep_shapes = TRUE) %>%
-    dplyr::group_by(geoid) %>%
-    sf::st_make_valid() %>%
-    # There are some duplicated geoids due to geomSmallCountries, but these are filtered out in the LoadCountry stage
-    dplyr::summarise(
-      m49code = dplyr::first(m49code),
-      iso3 = dplyr::first(iso3),
-      country_name = dplyr::first(country_name),
-      macro_code = dplyr::first(macro_code),
-      macro_name = dplyr::first(macro_name),
-      micro_code = dplyr::first(micro_code),
[…]
