+++ man/addNewGeoms.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/addNewGeom.R
+% Please edit documentation in R/geomUpdateUtils.R
 \name{addNewGeoms}
 \alias{addNewGeoms}
 \title{Add New Country Geometries}
@@ -10,9 +10,11 @@ addNewGeoms()
 world Map with new geoms added
 }
 \description{
-Add New Country Geometries
+Internal function for the geometry update process. Processes new geometries using "processX.R" scripts in the "tools/toProcess" folder and further formats the geometries correctly to be integrated into the system overall. Running this function outside of a geometry update may cause errors from moving process scripts prematurely.
 }
 \examples{
-addNewGeoms()
+\dontrun{
+localcovid19now:::addNewGeoms()
+}
 }
 \keyword{internal}

[…]
