### 2022-12-12 ee30d1f
make update geometry functions internal; fix issue that would show no errors when files were not copied (instead of overwriting) but would be deleted anyway; change messages to use `message` instead of `cat`
