+++ R/geomUpdateUtils.R
@@ -1,173 +1,12 @@
-#' Move Processed Files
-#'
-#' @description Moves files from "toProcess" to "processed" folder when updating country geometries. Designed to only work with the "processFileName" naming convention. This is part of the geometry updating system and is designed for developers.
-#' @param x File name to be moved
-#' @keywords internal
-#'
-#' @export
-#'
-#' @examples
-#' file.create(here::here("tools", "toProcess", "processTestFile.R"))
-#' moveProcessedFiles("processTestFile.R")
-moveProcessedFiles <- function(x) {
-  rlang::check_installed(c("here"), reason = "to use moveProcessedFiles()")
-  if (stringr::str_sub(x, 1, 7) != "process") {
-    cat("File", x, "is not formatted correctly. Please review standards.")
-  } else {
-    file.copy(
-      from = here::here("tools", "toProcess", x),
-      to = here::here("tools", "processed", x)
-    )
-    file.remove(here::here("tools", "toProcess", x))
-  }
-}
-
-#' Add New Country Geometries
-#'
-#' @description Internal function for the geometry update process. Processes new geometries using "processX.R" scripts in the "tools/toProcess" folder and further formats the geometries correctly to be integrated into the system overall. Running this function outside of a geometry update may cause errors from moving process scripts prematurely.
-#' @return world Map with new geoms added
-#' @keywords internal
-#' @export
-#'
-#' @examples
-#' addNewGeoms()
-addNewGeoms <- function() {
-  iso3 <- m49code <- macro_code <- micro_code <- geoid <- country_name <- macro_name <- NULL
-
-  rlang::check_installed(c("here"), reason = "to use addNewGeoms()")
-  # Fields will be as follows:
-  ## geoid = ISO3m49_micro_macro
-  ## m49code
-  ## iso3
-  ## county_name
-  ## macro_code
-  ## macro_name
-  ## micro_code
-  ## micro_name
-  ## geometry
-
-  necessary_cols <- dplyr::tibble("macro_code" = NA_character_, "macro_name" = NA_character_, "micro_code" = NA_character_, "micro_name" = NA_character_)
-
-  ## Load in the m49 codes for countries
-  utils::data("m49", envir = environment())
-
-  file_list <- list.files(here::here("tools", "toProcess"))[which(stringr::str_sub(list.files(here::here("tools", "toProcess")), 1, 7) == "process")] # only use files formatted as "processX"
-
-  if (length(file_list) > 0) {
-    # This "source" needs to be replaced bc it's bad form in a package, but I haven't figured out how it would work better yet
-    lapply(file_list, \(x) source(here::here("tools", "toProcess", x)))
-    geom_list <- paste0("geom", stringr::str_extract(file_list, "(?<=process)[:alpha:]+(?=.R$)"))
-
-    worldNew <- purrr::map_df(
-      geom_list,
-      ~ get(.x) %>%
-        dplyr::mutate(
-          filename = .x
-        ) %>%
-        sf::st_collection_extract() %>%
-        sf::st_cast("MULTIPOLYGON")
-    ) %>%
-      dplyr::left_join(necessary_cols) %>%
-      tidyr::replace_na(list("macro_code" = "00", "micro_code" = "00")) %>%
-      dplyr::mutate(
-        geoid = paste(paste0(iso3, m49code), macro_code, micro_code, sep = "_")
-      ) %>%
-      dplyr::select(geoid, m49code, iso3, country_name, macro_code, macro_name, dplyr::everything()) %>%
-      sf::st_as_sf() %>%
-      tibble::remove_rownames()
-
-    lapply(file_list, moveProcessedFiles)
-
-
-    return(worldNew)
-  } else {
-    cat("\nNo new geometries processed. toProcess folder is empty.\n")
-  }
-}
-
-#' Reset processed files
-#'
[…]
