+++ man/moveProcessedFiles.Rd
@@ -4,16 +4,18 @@
 \alias{moveProcessedFiles}
 \title{Move Processed Files}
 \usage{
-moveProcessedFiles(x)
+moveProcessedFiles(x, ...)
 }
 \arguments{
 \item{x}{File name to be moved}
+
+\item{...}{pass arguments to \code{file.copy}}
 }
 \description{
 Moves files from "toProcess" to "processed" folder when updating country geometries. Designed to only work with the "processFileName" naming convention. This is part of the geometry updating system and is designed for developers.
 }
 \examples{
 file.create(here::here("tools", "toProcess", "processTestFile.R"))
-moveProcessedFiles("processTestFile.R")
+loacalcovid19now:::moveProcessedFiles("processTestFile.R")
 }
 \keyword{internal}
