+++ man/addNewGeoms.Rd
@@ -13,6 +13,6 @@ world Map with new geoms added
 Internal function for the geometry update process. Processes new geometries using "processX.R" scripts in the "tools/toProcess" folder and further formats the geometries correctly to be integrated into the system overall. Running this function outside of a geometry update may cause errors from moving process scripts prematurely.
 }
 \examples{
-addNewGeoms()
+loacalcovid19now:::addNewGeoms()
 }
 \keyword{internal}
