+++ NAMESPACE
@@ -54,16 +54,11 @@ export(LoadVietnam)
 export(LoadZimbabwe)
 export(PerCapitaMap_leaflet)
 export(PerCapitaMap_tmap)
-export(addNewGeoms)
 export(assertOptions__tidy_Data)
 export(calcRisk)
 export(create_c19r_data)
 export(estRisk)
-export(moveProcessedFiles)
-export(remGeoSurplus)
-export(resetNewGeoms)
 export(tidy_Data)
-export(updateGlobal)
 importFrom(magrittr,"%>%")
 importFrom(rlang,":=")
 importFrom(rlang,check_installed)
