+++ man/LoadIndia.Rd
@@ -16,7 +16,9 @@ Reads in subnational data for India to calculate most recent estimate of per cap
 Data collated by \url{https://covid19tracker.in/}, an initiative of the Indian Institute of Technology Hyderabad, from state bulletins and official reports with thanks to the covid19india.org team for their outstanding work in creating the original portal, and for making their code base public.
 }
 \examples{
+\dontrun{
 India <- LoadIndia()
+}
 }
 \seealso{
 \code{\link[=LoadData]{LoadData()}}
