### 2022-12-13 48c4b72
update message formatting; and update googledrive scope checking in LoadData
