### 2022-12-13 2cce4f1
add description to check_gdrive_scope
