+++ man/check_gdrive_scope.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/LoadPhilippines.R
+\name{check_gdrive_scope}
+\alias{check_gdrive_scope}
+\title{check_gdrive_scope}
+\usage{
+check_gdrive_scope()
+}
+\description{
+runs a googledrive function that requires permissions to be shared by user with the tidyverse API. This function is designed to check if these permissions (needed for LoadPhilippines to work) are enabled, and error if not.
+}
+\examples{
+check_gdrive_scope()
+
+}
+\keyword{internal}
