### 2022-12-13 94a325e
update check_gdrive_scope
