+++ man/check_gdrive_scope.Rd
@@ -10,7 +10,8 @@ check_gdrive_scope()
 runs a googledrive function that requires permissions to be shared by user with the tidyverse API. This function is designed to check if these permissions (needed for LoadPhilippines to work) are enabled, and error if not.
 }
 \examples{
+\dontrun{
 check_gdrive_scope()
-
+}
 }
 \keyword{internal}
