+++ NAMESPACE
@@ -56,6 +56,7 @@ export(PerCapitaMap_leaflet)
 export(PerCapitaMap_tmap)
 export(assertOptions__tidy_Data)
 export(calcRisk)
+export(check_gdrive_scope)
 export(create_c19r_data)
 export(estRisk)
 export(tidy_Data)
