+++ R/LoadBrazil.R
@@ -11,7 +11,9 @@
 #' @return A simple feature returning the date of most recent data (DateReport), a unique region code (geoid), the region name (RegionName) and country name (Country), the number of active cases per capita (pInf) and the regions geometry (geometry).
 #'
 #' @examples
+#' \dontrun{
 #' Brazil <- LoadBrazil()
+#' }
 #' @seealso [LoadData()]
 #' @export
 LoadBrazil <- function() {
