### 2022-12-13 4c6bf6a
add Brazil and SwitzerlandLiechtenstein to dontrun as > 5s
move examples

update buildignore
