+++ README.md
@@ -62,7 +62,7 @@ PerCapitaMap_leaflet(GLOBALMAP,100000)
 PerCapitaMap_tmap(GLOBALMAP2,100000)
 ```
 
-![Active cases per capita across the globe](/examples/Global_pc_tmap.png)
+![Active cases per capita across the globe](/inst/examples/Global_pc_tmap.png)
 
 Additionally, with a provided assumed ascertainment bias (ratio between true infections and recorded cases), the risk that one or more people in a group of a particular size can be estimated and mapped:
 
@@ -86,7 +86,7 @@ tmap::tmap_save(MAP,"US_RiskMap.png")
 create_c19r_data(df_in = US, asc_bias_list = cbind(AB4 = US$AB))
 ```
 
-![event risk in the US](/examples/US_RiskMap.png)
+![event risk in the US](/inst/examples/US_RiskMap.png)
 
 We caution that the case ascertainment bias may differ both across different regions and across time (due to differences in testing strategies).
 
