+++ man/LoadSwitzerlandLiechtenstein.Rd
@@ -16,7 +16,9 @@ Reads in subnational data for Switzerland and also data for Liechtenstein to cal
 Data obtained via the Federal Office of Public Health FOPH \url{https://www.covid19.admin.ch/en/overview}.
 }
 \examples{
+\dontrun{
 SwitzerlandLiechtenstein <- LoadSwitzerlandLiechtenstein()
+}
 }
 \seealso{
 \code{\link[=LoadData]{LoadData()}}
