### 2022-12-13 1f9f128
Add map building to workflow
Signed-off-by: Aroon Chande <github@atc.io>
