+++ .github/workflows/check-release.yaml
@@ -2,9 +2,10 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, master]
+    branches: [main, master, workflow]
   pull_request:
     branches: [main, master]
+  workflow_dispatch:
 
 name: R-CMD-check
 
@@ -30,3 +31,39 @@ jobs:
         with:
           error-on: "'error'"
           upload-results: false
+  Make-Maps:
+    runs-on: ubuntu-latest
+    env:
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      R_KEEP_PKG_SOURCE: yes
+    steps:
+      - uses: actions/checkout@v3
+      - uses: r-lib/actions/setup-r@v2
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          extra-packages: local::., any::XML
+          needs: check
+
+      - name: Make global map
+        run: |
+          "png('world_map.png', width = 12, height = 8, res = 300, units = 'in'); localcovid19now::PerCapitaMap_leaflet(localcovid19now::LoadCountries(),100000); dev.off()"
+        shell: Rscript {0} 
+
+      - name: Upload images
+        uses: actions/upload-artifact@v3
+        with:
+          name: images
+          path: "*.png"
+
+      - name: Add map PNG links
+        uses: tonyhallett/artifacts-url-comments@v1.1.0
+        env:
+            GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+            prefix: Here are the map links
+            suffix: ''
+            format: url
+            addTo: pullandissues
+
+    
