+++ pkgdown.yml
@@ -1,7 +1,7 @@
-pandoc: 2.14.2
-pkgdown: 2.0.6
+pandoc: 2.19.2
+pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   updating-geometry-files: updating-geometry-files.html
-last_built: 2022-09-28T13:28Z
+last_built: 2022-12-15T18:37Z
 
