### 2022-12-15 de56348
Deploying to gh-pages from @ sjbeckett/localcovid19now@e53411d1708a6c7c97ff3aaecf0f674018739ce2 🚀
