### 2022-12-15 a04c372
Limit maps workflow to PRs
Signed-off-by: Aroon Chande <github@atc.io>
