+++ .github/workflows/make_maps.yml
@@ -3,8 +3,6 @@
 name: maps
 
 on:
-  push:
-    branches: [main, master ]
   pull_request:
     branches: [main, master, revise]
   workflow_dispatch:
