+++ .github/workflows/pr-links.yaml
@@ -1,0 +1,24 @@
+
+name: Add artifact link to pull request
+
+on:
+  workflow_run:
+    workflows: 
+      - maps
+    types: 
+      - completed
+  workflow_dispatch:
+
+jobs:
+  artifacts-url-comments:
+    name: Add artifact link to pull request
+    runs-on: ubuntu-latest
+    steps:
+      - name: Add artifact link to pull request
+        uses: tonyhallett/artifacts-url-comments@v1.1.0
+        env:
+            GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+            prefix: "Check images for correctness:"
+            format: names
+            addTo: pull
