### 2022-12-15 2a23bfa
Add map building to workflow (#64)
* Add pr-links action
* Update pr-links.yaml
* Limit maps workflow to PRs
