### 2022-12-15 43600ed
Add map building to workflow (#64)
* Add pr-links action
* Update pr-links.yaml
* Limit maps workflow to PRs

Filter to R and data changes only

Signed-off-by: Aroon Chande <github@atc.io>
