+++ .github/workflows/make_maps.yml
@@ -1,0 +1,68 @@
+# Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
+# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
+name: maps
+
+on:
+  pull_request:
+    branches: [main, master, revise]
+    paths: 
+      - 'R/*.R'
+      - 'data/*'
+
+  workflow_dispatch:
+
+
+jobs:
+  Make-Maps:
+    runs-on: ubuntu-latest
+    env:
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      R_KEEP_PKG_SOURCE: yes
+    steps:
+      - uses: actions/checkout@v3
+      - uses: r-lib/actions/setup-r@v2
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          extra-packages: local::., any::XML
+
+      - name: Make global map
+        run: |
+          echo "library(localcovid19now)" > worldmap.R
+          echo "png('./world_map.png', width = 12, height = 8, res = 300, units = 'in')" > worldmap.R
+          echo "localcovid19now::PerCapitaMap_tmap(localcovid19now::LoadData(), 100000) " >> worldmap.R
+          echo "dev.off()" >> worldmap.R
+          Rscript worldmap.R
+
+      - name: List files
+        run: |
+            ls -lthr
+
+      - uses: actions/upload-artifact@v3
+        with:
+          name: world_map.png
+          path: world_map.png
+
+
+      # - name: Run Examples
+      #   run: |
+      #     Rscript  "inst/examples/examples_from_README.R"
+
+      # - uses: actions/upload-artifact@v3
+      #   with:
+      #     name: USrisk.csv
+      #     path: USrisk.csv
+
+      # - uses: actions/upload-artifact@v3
+      #   with:
+      #     name: US_RiskMap.png
+      #     path: US_RiskMap.png
+
+      # - uses: actions/upload-artifact@v3
+      #   with:
+      #     name: Global_pc_tmap.png
+      #     path: Global_pc_tmap.png
+
+
+
+
