### 2022-12-15 0d37e0f
- add tests
- add extra check on RiskEval options

- add note on Belgium checking multiple files
