+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(localcovid19now)
+
+test_check("localcovid19now")
