+++ man/LoadBelgium.Rd
@@ -13,7 +13,7 @@ A simple feature returning the date of most recent data (DateReport), a unique r
 Reads in subnational data for Belgium to calculate most recent estimate of per capita active COVID-19 cases. Use with LoadData() is recommended.
 }
 \note{
-Data was obtained from Sciensano, the Belgian institute for health: \url{https://epistat.wiv-isp.be/covid/}
+Data was obtained from Sciensano, the Belgian institute for health: \url{https://epistat.wiv-isp.be/covid/}. We note that this function needs to read files corresponding to daily reports (some of which may not exist),
 }
 \examples{
 \dontrun{
