### 2022-12-15 87ed42d
- add tests
- fix typo in updateGlobal() comments
- fix Fiji in the geomWorld_presimp file on record, which fixes an error when comparing existing and "new" geomSmallCountries
- update version
