+++ man/localcovid19now-package.Rd
@@ -6,7 +6,7 @@
 \alias{localcovid19now-package}
 \title{localcovid19now: Processing and mapping COVID-19 case data at subnational scales}
 \description{
-Extract, clean and map recent subnational level data on active COVID-19 cases across multiple countries. Provides tools to extract and clean data from multiple data sources; and tools to visualize, via mapping, the data in terms of per capita active infections; or as the risk (%) that one or more individuals in a group of a particular size are infectious.
+Extract, clean and map recent subnational level data on active COVID-19 cases across multiple countries. Provides tools to extract and clean data from multiple data sources; and tools to visualize, via mapping, the data in terms of per capita active infections; or as the risk that one or more individuals in a group of a particular size are infectious.
 }
 \seealso{
 Useful links:
@@ -26,7 +26,12 @@ Authors:
   \item Aroon T. Chande \email{achande@ihrc.com} (\href{https://orcid.org/0000-0001-5260-3135}{ORCID})
   \item Lavanya Rishishwar \email{lrishishwar@ihrc.com} (\href{https://orcid.org/0000-0002-2055-9392 }{ORCID})
   \item Clio Andris \email{clio@gatech.edu} (\href{https://orcid.org/0000-0002-8559-5079}{ORCID})
-  \item Joshua S. Weitz \email{jsweitz@gatech.edu} (\href{https://orcid.org/0000-0002-3433-8312 }{ORCID})
+  \item Joshua S. Weitz \email{jsweitz@gatech.edu} (\href{https://orcid.org/0000-0002-3433-8312}{ORCID})
+}
+
+Other contributors:
+\itemize{
+  \item Kaitlyn Johnson \email{kjohnson@rockfound.org} (\href{https://orcid.org/0000-0001-8011-0012}{ORCID}) [contributor]
 }
 
 }

[…]
