### 2022-12-15 ed4e54f
add extra check on RiskEval options
