### 2022-12-15 b4ffd47
- fix typo in updateGlobal() comments - fix Fiji in the geomWorld_presimp file on record, which fixes an error when comparing existing and "new" geomSmallCountries - update description to remove % - update FAB
