### 2022-12-16 #4898 @galessiorob
Checking in 🎉

@welch16 and @mbkumar thanks again for the review and the improvement recommendations!
@welch16 I agree that including unit tests would be a good addition on top of GH Actions.

@sjbeckett please let me know when this is ready for another full review - I hope we can this paper published soon!
