### 2022-12-16 #4898 @sjbeckett
Many thanks @galessiorob. We are getting close and plan to merge revisions into our main branch and send a response early next week.

We have added a github actions check, and have implemented unit tests for some of our core functions, though we think unit tests would be tricky to implement across the loading functions as we noted in response to @welch16 https://github.com/sjbeckett/localcovid19now/issues/62#issuecomment-1353344286 . More soon!
