### 2022-12-18 8b75358
Filter to R and data changes only
Signed-off-by: Aroon Chande <github@atc.io>
