+++ paper/paper.md
@@ -57,7 +57,7 @@ The *[localcovid19now](https://github.com/sjbeckett/localcovid19now)* R package 
 The COVID-19 global pandemic has led to hundreds of millions of recorded cases, over six million recorded deaths, and critical pressures on both health and economic systems [@10665-362241]. A key challenge in mitigating ongoing spread is that disease transmission and human behavior are coupled [@funk2010modelling; @bavel2020using; @weitz2020awareness]. Coupled viral and human dynamics have limited the predictive window of infectious disease models (including those of COVID-19), while also catalyzing efforts to develop awareness campaigns that leverage localized real-time information to inform the public and mitigate transmission. Beginning in May 2020, our team launched the COVID-19 Event Risk Tool (<https://covid19risk.biosci.gatech.edu/>), a situational dashboard with maps estimating exposure risk to COVID-19 across counties in the US [@chande2020real]. We use the number of active cases per capita and an assumed case ascertainment bias to estimate current local COVID-19 prevalence $p$. We then quantify and visualize COVID-19 exposure risk $\kappa$ (%) within a group of size $n$ as $\kappa = 100\times(1-(1-p)^n)$. Since launch, more than 16M individuals have accessed the site and several similar international efforts have been initiated, e.g., for [Italy](https://covid19eventi.datainterfaces.org/) [@Italy], [Spain](https://eventosycovid19.es/) [@Spain], [Mexico](https://adrian-acuna.shinyapps.io/Dashboard-MX-v3/_w_33ffcba1/#shiny-tab-Riesgo) [@Mexico], [Argentina](https://www.lanacion.com.ar/sociedad/fiestas-fin-ano-calcula-riesgo-coronavirus-infectado-nid2548498/#/) [@Argentina]. Here, in the *localcovid19now* R package we extend our initial release [@chande2020real] to develop a pipeline to combine population estimates, COVID-19 case data, and spatial geometry such that risk estimates can be performed at localized scales across multiple countries. Our package aims to facilitate the process of downloading and visualizing recent COVID-19 data, in support of public health research and campaigns to inform and alert the public of the ongoing risk of transmission -- and is of particular interest to public health scientists and researchers, epidemiologists, community leaders, and policymakers.
 
[…]
