### 2022-12-22 sjbeckett/localcovid19now#63 @sjbeckett
Edited in with #67
