### 2022-12-22 sjbeckett/localcovid19now#62 @sjbeckett
Unit tests for some core functions added with #67
