### 2022-12-22 sjbeckett/localcovid19now#59 @sjbeckett
This may be due to changes in support in underlying geometry formalism (e.g. see https://github.com/r-spatial/sf/issues/1419). We have now updated the examples in #67.
