### 2022-12-22 #4898 @sjbeckett
Many thanks to all of you for taking the time to review our package and provide critical feedback -- and thank you @galessiorob for handling the review process. We have spent time working through your feedback and suggestions which we outline below – these edits are now committed and merged into the main branch of sjbeckett/localcovid19now which we have released as version 0.1. A link to the pull request with changes is here: [sjbeckett/localcovid19now/pull/67](https://github.com/sjbeckett/localcovid19now/pull/67). Our revisions are as follows:

Manuscript:

1. We now more explicitly outline our intended audience by including the statement:

“Our package aims to facilitate the process of downloading and visualizing recent COVID-19 data, in support of public health research and campaigns to inform and alert the public of the ongoing risk of transmission -- and is of particular interest to public health scientists and researchers, epidemiologists, community leaders, and policymakers.“

2. We updated reference formatting as suggested by @Epic19mz.

3. We now incorporate a figure showing an example output as suggested by @richelbilderbeek. We purposefully chose to show a figure that represents the scope of the project, including regions which are not currently reporting recent data (an issue described in the manuscript).

4. To clarify author contributions we now include a statement in the manuscript. Nguyen's contributions were predominantly pre-GitHub, with contributions to overall data collection methodology, as well as coding up loading functions for several countries including, but not limited to Belgium, Brazil, China, Netherlands, Norway, Japan, Spain. We also note that important contributions came from Rishishwar, Andris and Weitz who provided technical guidance and project management. We include the following author attribution statement:

“SJB led the project. SJB, FB-T, QN, and ATC contributed to dataset collection and coding methods for data curation. SJB, FB-T, and ATC contributed to package development. LR, CA, and JSW provided substantial contributions via decision and design guidance, and project management. SJB wrote the manuscript with contributions from all authors.”

Reviewers noted that Johnson has provided contributions to our project. We appreciate Johnson’s contribution to patching our `LoadCanada` function when the location of the datasource changed. Johnson is acknowledged in our manuscript, and we have added Johnson as a contributor “ctb” within our DESCRIPTION file, following: https://stat.ethz.ch/R-manual/R-devel/library/utils/html/person.html.

README and Examples:

1. We have updated the example documentation in the README.md and corresponding examples directory as noted by @mbkumar here: [sjbeckett/localcovid19now/issues#59](https://github.com/sjbeckett/localcovid19now/issues/59)

2. We also provided more information about using the `googledrive` library (needed for `LoadPhilippines`) as suggested by @richelbilderbeek, which will not work unless certain permissions are accepted by the user – even if a `googledrive` token is acquired! We added a check within `LoadPhilippines` to check if users have supplied permissions such that the function can run via `googledrive::drive_about()`, and returned a more user-friendly error message (than the gargle permissions error) with instructions on how users can update permissions and authorization by running `googledrive::drive_auth()`.

3. We noted that users may need to load up packages that our package imports from as suggested by @mbkumar. Hopefully this will be seen as more user friendly – and we do sympathize as we have also previously faced these issues using R – but, as these challenges in installation are rooted in other packages, we believe this goes beyond the scope of our package. We note for completeness that the packages within the R ecosystem that are used within our package are listed in the DESCRIPTION file; which is used by R’s package management system to automate installation/loading.

4. We also added a note regarding the possibility of changing timeout options in helping users to install the package as found by @welch16 in the issue raised:[ sjbeckett/localcovid19now#58](https://github.com/sjbeckett/localcovid19now/issues/58)

Code Base:

1. We fixed data source file loading for New Zealand and Italy, which were previously failing.

2. We suppressed progress, summary, and console out messages within data loading calls (and `create_c19r_data`) following advice from @richelbilderbeek. While we agree some of the prior summary output (via the `vroom` package) is nice as noted by @Epic19mz we preferred this approach as it reduces verbosity and provides more cohesion across different data loading calls (not all data is loaded via `vroom` due to differences in underlying file types/storage).

3. We updated package documentation (including `PerCapitaMap_tmap`, `PerCapitaMap_leaflet`, `EventMap_tmap`, `EventMap_leaflet`) inline with suggestions from @Epic19mz and @richelbilderbeek and removed many instances of dontrun across the package. Some instances of dontrun remain, as some functions take >5s to run.

4. R CMD checks are run on submitting pull requests, and we have worked to fix and remove warnings and notes that were previously generated by these checks by removing .gitkeep files, and moving nonstandard files/folders into the inst folder or adding them to the .Rbuildignore list.

5. We revisited how tidy_Data was being applied within the package. We now encourage users to use the function `LoadData`, which performs both loading and tidying of the dataset (using the `tidy_Data`). i.e. we now instruct users to run `LoadData(“LoadUS”)` rather than `LoadUS()`. We have updated examples in the README to reflect this. Additionally to help usability, we have included some alternative matchings to help users i.e. the calls `LoadData(“USA”)`, `LoadData(“LoadUS”)` and `LoadData(“United States”)` all produce equivalent output. A list of main function calls are stored in `countrylist`, with a list of alternate matchings stored in `alt_countrylist`.

6. We added checking of user inputs within `LoadData`, such that user inputs are of the correct type and contain appropriate values, with informative error messaging if not, which should improve the usability of our package.

7. We note that some functionality within our package is geared more towards development purposes than for a typical user. At the same time, we believe these files are more useful if they are kept with the package than removed/moved elsewhere. As such, we have made the functions previously within files: `addnewGeom.R`,`updateGlobal.R`, `remSurplus.R` internalised functions, and moved them together into a new file geomUpdateUtils.R. Additionally, as suggested by @richelbilderbeek we renamed the function `moveFiles` to `moveProcessedFiles`.

Automated tests:

1. Following @richelbilderbeek’s suggestion we introduced a new github action check (in addition to R CMD checks) to test package functionality by attempting to download all data files and render them via PerCaptiaMap_tmap following a pull request.

2. @welch16 had a great recommendation for incorporating unit tests. We chose to implement unit tests for some common utility functions used by our package, namely `tidy_Data`, `assertOptions__tidy_data`, `estRisk` and `calcRisk`. However, given the dynamic nature of the data the loading functions ingest, the output is non-deterministic which makes unit testing significantly harder to implement across the entire package. Alternatively, we could break each function into smaller units but that would entail a complete rewrite of the code to accommodate unit tests.  In lieu of this, we have added additional failsafes to improve user experience and make the package more rigorous.  In particular, we have implemented `tryCatch` loops within `LoadData` to better handle any potential errors from any of the particular data source loads. We note our expectation is that many of the underlying datasets are not going to be stable resources in the long-term (which is ok!).

Other comments:

1. We note to @Epic19mz that our focus in the package is on recent data, but some data resources are saved using the entire timeseries (e.g. `LoadAustralia`), while others have different files corresponding to data on individual days (e.g. `LoadBelgium`). Such differences were a key motivator in developing this project.

2. We additionally updated code for Canada following changes in source highlighted in a new issue outside of the context of this review: [sjbeckett/localcovid19now#63](https://github.com/sjbeckett/localcovid19now/issues/63)

3. There are some comments we have not been able to fully resolve. In particular, while we are able to run `LoadAustria` and `LoadDenmark` locally, they seem to not run within the automated ubuntu environment we set up to generate maps. We have also been unable to find a solution to prevent warning errors regarding closing unused connections from appearing in the console. These errors correspond to attempting to open invalid URL’s as we search for recent data (date is encoded in the URL string) within the `LoadBelgium` and `LoadMexico` functions.

Again, thank you all for your feedback leading to these revisions. Please reach out if you have further questions.
