### 2022-12-23 sjbeckett/localcovid19now#63 @jeanpaulrsoucy
Great work, @sjbeckett. One last thing:

https://github.com/sjbeckett/localcovid19now/blob/3de9acf76aec816ecb1c706d4977273d5792d682/tools/processed/processCanada.R#L2

Should be:

```
geomCountry <- sf::st_read("https://raw.githubusercontent.com/ccodwg/CovidTimelineCanada/main/geo/hr.geojson")
```

Thanks!
