### 2022-12-23 3da40c9
Update geojson based on #53
Signed-off-by: Aroon Chande <github@atc.io>
