### 2022-12-23 134ad97
Glob all in R and data for map_maps workflow
