### 2022-12-23 4a343e1
Deploying to gh-pages from @ sjbeckett/localcovid19now@3da40c9830f3a3919b25400a0a3cb00b8a55794b 🚀
