### 2023-01-09 #4898 @sjbeckett
> @sjbeckett amazing work! I just did a final run looking at the latest changes, and all looks good 🚀
>
> I put together a few suggestions for the actual paper in [this branch](https://github.com/galessiorob/localcovid19now/blob/galessiorob-edit-suggestions/paper/paper.md) from a fork. They are minor, please incorporate them as you see fit.
>
> After that, I think we should be ready to publish.

Many thanks @galessiorob. I have incorporated your suggestions for the manuscript this morning. Excited to move forward!
