### 2023-01-16 #4898 @galessiorob
I think we're ready for the final steps @sjbeckett!

@mbkumar could you please take a look at the final paper and the last updates to the code and check all remaining items on your list if you are satisfied?

@Epic19mz I checked off yours based on [your comment](https://github.com/openjournals/joss-reviews/issues/4898#issuecomment-1371270549) ✅ . I also removed your duplicated handle (`@epic19mz`.)

@sjbeckett the next thing we need from you is to deposit your software in Zenodo and post the DOI here and the most recent version. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing. 🙏
