### 2023-01-21 #4898 @Kevin-Mattheus-Moerman
@sjbeckett I am the AEiC for this track and here to help process the final steps towards acceptance in JOSS. The archive and paper look good. On the paper's affiliations, can you please spell out USA as United States of America for each? Thanks.
